Move Composite from common.ts to composites/composite.ts.
