+++ docs/api/assets/search.js
@@ -1,1 +1,1 @@
-window.searchData = JSON.parse("{\"kinds\":{\"64\":\"Function\",\"128\":\"Class\",\"256\":\"Interface\",\"512\":\"Constructor\",\"1024\":\"Property\",\"2048\":\"Method\",\"65536\":\"Type literal\",\"262144\":\"Accessor\",\"4194304\":\"Type alias\",\"8388608\":\"Reference\"},\"rows\":[{\"kind\":128,\"name\":\"CandyGraph\",\"url\":\"classes/CandyGraph.html\",\"classes\":\"tsd-kind-class\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/CandyGraph.html#constructor\",\"classes\":\"tsd-kind-constructor tsd-parent-kind-class\",\"parent\":\"CandyGraph\"},{\"kind\":1024,\"name\":\"canvas\",\"url\":\"classes/CandyGraph.html#canvas\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"CandyGraph\"},{\"kind\":1024,\"name\":\"commandsByCoords\",\"url\":\"classes/CandyGraph.html#commandsByCoords\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"CandyGraph\"},{\"kind\":1024,\"name\":\"scope\",\"url\":\"classes/CandyGraph.html#scope\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"CandyGraph\"},{\"kind\":2048,\"name\":\"clear\",\"url\":\"classes/CandyGraph.html#clear\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"CandyGraph\"},{\"kind\":2048,\"name\":\"render\",\"url\":\"classes/CandyGraph.html#render\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"CandyGraph\"},{\"kind\":2048,\"name\":\"destroy\",\"url\":\"classes/CandyGraph.html#destroy\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"CandyGraph\"},{\"kind\":2048,\"name\":\"copyTo\",\"url\":\"classes/CandyGraph.html#copyTo\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"CandyGraph\"},{\"kind\":2048,\"name\":\"getNamedCommands\",\"url\":\"classes/CandyGraph.html#getNamedCommands\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-private\",\"parent\":\"CandyGraph\"},{\"kind\":2048,\"name\":\"recursiveRender\",\"url\":\"classes/CandyGraph.html#recursiveRender\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-private\",\"parent\":\"CandyGraph\"},{\"kind\":256,\"name\":\"CandyGraphOptions\",\"url\":\"interfaces/CandyGraphOptions.html\",\"classes\":\"tsd-kind-interface\"},{\"kind\":1024,\"name\":\"canvas\",\"url\":\"interfaces/CandyGraphOptions.html#canvas\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"CandyGraphOptions\"},{\"kind\":1024,\"name\":\"alpha\",\"url\":\"interfaces/CandyGraphOptions.html#alpha\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"CandyGraphOptions\"},{\"kind\":256,\"name\":\"Viewport\",\"url\":\"interfaces/Viewport.html\",\"classes\":\"tsd-kind-interface\"},{\"kind\":1024,\"name\":\"x\",\"url\":\"interfaces/Viewport.html#x\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"Viewport\"},{\"kind\":1024,\"name\":\"y\",\"url\":\"interfaces/Viewport.html#y\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"Viewport\"},{\"kind\":1024,\"name\":\"width\",\"url\":\"interfaces/Viewport.html#width\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"Viewport\"},{\"kind\":1024,\"name\":\"height\",\"url\":\"interfaces/Viewport.html#height\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"Viewport\"},{\"kind\":128,\"name\":\"Composite\",\"url\":\"classes/Composite.html\",\"classes\":\"tsd-kind-class\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/Composite.html#constructor\",\"classes\":\"tsd-kind-constructor tsd-parent-kind-class\",\"parent\":\"Composite\"},{\"kind\":2048,\"name\":\"dispose\",\"url\":\"classes/Composite.html#dispose\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"Composite\"},{\"kind\":4194304,\"name\":\"Renderable\",\"url\":\"types/Renderable.html\",\"classes\":\"tsd-kind-type-alias\"},{\"kind\":4194304,\"name\":\"NumberArray\",\"url\":\"types/NumberArray.html\",\"classes\":\"tsd-kind-type-alias\"},{\"kind\":4194304,\"name\":\"Vector2\",\"url\":\"types/Vector2.html\",\"classes\":\"tsd-kind-type-alias\"},{\"kind\":4194304,\"name\":\"Vector3\",\"url\":\"types/Vector3.html\",\"classes\":\"tsd-kind-type-alias\"},{\"kind\":4194304,\"name\":\"Vector4\",\"url\":\"types/Vector4.html\",\"classes\":\"tsd-kind-type-alias\"},{\"kind\":4194304,\"name\":\"NamedDrawCommands\",\"url\":\"types/NamedDrawCommands.html\",\"classes\":\"tsd-kind-type-alias\"},{\"kind\":128,\"name\":\"Primitive\",\"url\":\"classes/Primitive.html\",\"classes\":\"tsd-kind-class\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/Primitive.html#constructor\",\"classes\":\"tsd-kind-constructor tsd-parent-kind-class\",\"parent\":\"Primitive\"},{\"kind\":2048,\"name\":\"dispose\",\"url\":\"classes/Primitive.html#dispose\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"Primitive\"},{\"kind\":128,\"name\":\"Dataset\",\"url\":\"classes/Dataset.html\",\"classes\":\"tsd-kind-class\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/Dataset.html#constructor\",\"classes\":\"tsd-kind-constructor tsd-parent-kind-class\",\"parent\":\"Dataset\"},{\"kind\":1024,\"name\":\"length\",\"url\":\"classes/Dataset.html#length\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"Dataset\"},{\"kind\":1024,\"name\":\"disposed\",\"url\":\"classes/Dataset.html#disposed\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"Dataset\"},{\"kind\":2048,\"name\":\"update\",\"url\":\"classes/Dataset.html#update\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"Dataset\"},{\"kind\":2048,\"name\":\"dispose\",\"url\":\"classes/Dataset.html#dispose\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"Dataset\"},{\"kind\":2048,\"name\":\"assertFits\",\"url\":\"classes/Dataset.html#assertFits\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-private\",\"parent\":\"Dataset\"},{\"kind\":128,\"name\":\"LinearScale\",\"url\":\"classes/LinearScale.html\",\"classes\":\"tsd-kind-class\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/LinearScale.html#constructor\",\"classes\":\"tsd-kind-constructor tsd-parent-kind-class\",\"parent\":\"LinearScale\"},{\"kind\":1024,\"name\":\"_domain\",\"url\":\"classes/LinearScale.html#_domain\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"LinearScale\"},{\"kind\":1024,\"name\":\"_range\",\"url\":\"classes/LinearScale.html#_range\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"LinearScale\"},{\"kind\":1024,\"name\":\"qDomain\",\"url\":\"classes/LinearScale.html#qDomain\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"LinearScale\"},{\"kind\":1024,\"name\":\"qRange\",\"url\":\"classes/LinearScale.html#qRange\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"LinearScale\"},{\"kind\":262144,\"name\":\"domain\",\"url\":\"classes/LinearScale.html#domain\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class\",\"parent\":\"LinearScale\"},{\"kind\":262144,\"name\":\"range\",\"url\":\"classes/LinearScale.html#range\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class\",\"parent\":\"LinearScale\"},{\"kind\":2048,\"name\":\"toDomain\",\"url\":\"classes/LinearScale.html#toDomain\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"LinearScale\"},{\"kind\":2048,\"name\":\"toRange\",\"url\":\"classes/LinearScale.html#toRange\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"LinearScale\"},{\"kind\":2048,\"name\":\"updateQ\",\"url\":\"classes/LinearScale.html#updateQ\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-private\",\"parent\":\"LinearScale\"},{\"kind\":128,\"name\":\"LogScale\",\"url\":\"classes/LogScale.html\",\"classes\":\"tsd-kind-class\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/LogScale.html#constructor\",\"classes\":\"tsd-kind-constructor tsd-parent-kind-class\",\"parent\":\"LogScale\"},{\"kind\":1024,\"name\":\"_domain\",\"url\":\"classes/LogScale.html#_domain\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"LogScale\"},{\"kind\":1024,\"name\":\"_range\",\"url\":\"classes/LogScale.html#_range\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"LogScale\"},{\"kind\":1024,\"name\":\"conversion\",\"url\":\"classes/LogScale.html#conversion\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"LogScale\"},{\"kind\":1024,\"name\":\"base\",\"url\":\"classes/LogScale.html#base\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"LogScale\"},{\"kind\":262144,\"name\":\"domain\",\"url\":\"classes/LogScale.html#domain\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class\",\"parent\":\"LogScale\"},{\"kind\":262144,\"name\":\"range\",\"url\":\"classes/LogScale.html#range\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class\",\"parent\":\"LogScale\"},{\"kind\":2048,\"name\":\"toRange\",\"url\":\"classes/LogScale.html#toRange\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"LogScale\"},{\"kind\":2048,\"name\":\"toDomain\",\"url\":\"classes/LogScale.html#toDomain\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"LogScale\"},{\"kind\":128,\"name\":\"CartesianCoordinateSystem\",\"url\":\"classes/CartesianCoordinateSystem.html\",\"classes\":\"tsd-kind-class\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/CartesianCoordinateSystem.html#constructor\",\"classes\":\"tsd-kind-constructor tsd-parent-kind-class\",\"parent\":\"CartesianCoordinateSystem\"},{\"kind\":1024,\"name\":\"xscale\",\"url\":\"classes/CartesianCoordinateSystem.html#xscale\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"CartesianCoordinateSystem\"},{\"kind\":1024,\"name\":\"yscale\",\"url\":\"classes/CartesianCoordinateSystem.html#yscale\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"CartesianCoordinateSystem\"},{\"kind\":2048,\"name\":\"toRange\",\"url\":\"classes/CartesianCoordinateSystem.html#toRange\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"CartesianCoordinateSystem\"},{\"kind\":2048,\"name\":\"toDomain\",\"url\":\"classes/CartesianCoordinateSystem.html#toDomain\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"CartesianCoordinateSystem\"},{\"kind\":128,\"name\":\"PolarCoordinateSystem\",\"url\":\"classes/PolarCoordinateSystem.html\",\"classes\":\"tsd-kind-class\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/PolarCoordinateSystem.html#constructor\",\"classes\":\"tsd-kind-constructor tsd-parent-kind-class\",\"parent\":\"PolarCoordinateSystem\"},{\"kind\":1024,\"name\":\"radialScale\",\"url\":\"classes/PolarCoordinateSystem.html#radialScale\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"PolarCoordinateSystem\"},{\"kind\":1024,\"name\":\"angularScale\",\"url\":\"classes/PolarCoordinateSystem.html#angularScale\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"PolarCoordinateSystem\"},{\"kind\":1024,\"name\":\"xScale\",\"url\":\"classes/PolarCoordinateSystem.html#xScale\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"PolarCoordinateSystem\"},{\"kind\":1024,\"name\":\"yScale\",\"url\":\"classes/PolarCoordinateSystem.html#yScale\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"PolarCoordinateSystem\"},{\"kind\":2048,\"name\":\"toRange\",\"url\":\"classes/PolarCoordinateSystem.html#toRange\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"PolarCoordinateSystem\"},{\"kind\":2048,\"name\":\"toDomain\",\"url\":\"classes/PolarCoordinateSystem.html#toDomain\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"PolarCoordinateSystem\"},{\"kind\":128,\"name\":\"CoordinateSystem\",\"url\":\"classes/CoordinateSystem.html\",\"classes\":\"tsd-kind-class\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/CoordinateSystem.html#constructor\",\"classes\":\"tsd-kind-constructor tsd-parent-kind-class\",\"parent\":\"CoordinateSystem\"},{\"kind\":2048,\"name\":\"toDomain\",\"url\":\"classes/CoordinateSystem.html#toDomain\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"CoordinateSystem\"},{\"kind\":2048,\"name\":\"toRange\",\"url\":\"classes/CoordinateSystem.html#toRange\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"CoordinateSystem\"},{\"kind\":128,\"name\":\"Circles\",\"url\":\"classes/Circles.html\",\"classes\":\"tsd-kind-class\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/Circles.html#constructor\",\"classes\":\"tsd-kind-constructor tsd-parent-kind-class\",\"parent\":\"Circles\"},{\"kind\":1024,\"name\":\"xs\",\"url\":\"classes/Circles.html#xs\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Circles\"},{\"kind\":1024,\"name\":\"ys\",\"url\":\"classes/Circles.html#ys\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Circles\"},{\"kind\":1024,\"name\":\"colors\",\"url\":\"classes/Circles.html#colors\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Circles\"},{\"kind\":1024,\"name\":\"radii\",\"url\":\"classes/Circles.html#radii\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Circles\"},{\"kind\":1024,\"name\":\"borderWidths\",\"url\":\"classes/Circles.html#borderWidths\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Circles\"},{\"kind\":1024,\"name\":\"borderColors\",\"url\":\"classes/Circles.html#borderColors\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Circles\"},{\"kind\":1024,\"name\":\"positionBuffer\",\"url\":\"classes/Circles.html#positionBuffer\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"Circles\"},{\"kind\":2048,\"name\":\"dispose\",\"url\":\"classes/Circles.html#dispose\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"Circles\"},{\"kind\":256,\"name\":\"CirclesOptions\",\"url\":\"interfaces/CirclesOptions.html\",\"classes\":\"tsd-kind-interface\"},{\"kind\":1024,\"name\":\"colors\",\"url\":\"interfaces/CirclesOptions.html#colors\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"CirclesOptions\"},{\"kind\":1024,\"name\":\"radii\",\"url\":\"interfaces/CirclesOptions.html#radii\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"CirclesOptions\"},{\"kind\":1024,\"name\":\"borderColors\",\"url\":\"interfaces/CirclesOptions.html#borderColors\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"CirclesOptions\"},{\"kind\":1024,\"name\":\"borderWidths\",\"url\":\"interfaces/CirclesOptions.html#borderWidths\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"CirclesOptions\"},{\"kind\":128,\"name\":\"HLines\",\"url\":\"classes/HLines.html\",\"classes\":\"tsd-kind-class\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/HLines.html#constructor\",\"classes\":\"tsd-kind-constructor tsd-parent-kind-class\",\"parent\":\"HLines\"},{\"kind\":1024,\"name\":\"lines\",\"url\":\"classes/HLines.html#lines\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"HLines\"},{\"kind\":1024,\"name\":\"widths\",\"url\":\"classes/HLines.html#widths\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"HLines\"},{\"kind\":1024,\"name\":\"colors\",\"url\":\"classes/HLines.html#colors\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"HLines\"},{\"kind\":1024,\"name\":\"segmentGeometry\",\"url\":\"classes/HLines.html#segmentGeometry\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"HLines\"},{\"kind\":2048,\"name\":\"dispose\",\"url\":\"classes/HLines.html#dispose\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"HLines\"},{\"kind\":256,\"name\":\"HLinesOptions\",\"url\":\"interfaces/HLinesOptions.html\",\"classes\":\"tsd-kind-interface\"},{\"kind\":1024,\"name\":\"widths\",\"url\":\"interfaces/HLinesOptions.html#widths\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"HLinesOptions\"},{\"kind\":1024,\"name\":\"colors\",\"url\":\"interfaces/HLinesOptions.html#colors\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"HLinesOptions\"},{\"kind\":128,\"name\":\"InterleavedCircles\",\"url\":\"classes/InterleavedCircles.html\",\"classes\":\"tsd-kind-class\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/InterleavedCircles.html#constructor\",\"classes\":\"tsd-kind-constructor tsd-parent-kind-class\",\"parent\":\"InterleavedCircles\"},{\"kind\":1024,\"name\":\"xys\",\"url\":\"classes/InterleavedCircles.html#xys\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"InterleavedCircles\"},{\"kind\":1024,\"name\":\"colors\",\"url\":\"classes/InterleavedCircles.html#colors\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"InterleavedCircles\"},{\"kind\":1024,\"name\":\"radii\",\"url\":\"classes/InterleavedCircles.html#radii\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"InterleavedCircles\"},{\"kind\":1024,\"name\":\"borderWidths\",\"url\":\"classes/InterleavedCircles.html#borderWidths\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"InterleavedCircles\"},{\"kind\":1024,\"name\":\"borderColors\",\"url\":\"classes/InterleavedCircles.html#borderColors\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"InterleavedCircles\"},{\"kind\":1024,\"name\":\"positionBuffer\",\"url\":\"classes/InterleavedCircles.html#positionBuffer\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"InterleavedCircles\"},{\"kind\":2048,\"name\":\"dispose\",\"url\":\"classes/InterleavedCircles.html#dispose\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"InterleavedCircles\"},{\"kind\":256,\"name\":\"InterleavedCirclesOptions\",\"url\":\"interfaces/InterleavedCirclesOptions.html\",\"classes\":\"tsd-kind-interface\"},{\"kind\":1024,\"name\":\"colors\",\"url\":\"interfaces/InterleavedCirclesOptions.html#colors\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"InterleavedCirclesOptions\"},{\"kind\":1024,\"name\":\"borderColors\",\"url\":\"interfaces/InterleavedCirclesOptions.html#borderColors\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"InterleavedCirclesOptions\"},{\"kind\":1024,\"name\":\"radii\",\"url\":\"interfaces/InterleavedCirclesOptions.html#radii\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"InterleavedCirclesOptions\"},{\"kind\":1024,\"name\":\"borderWidths\",\"url\":\"interfaces/InterleavedCirclesOptions.html#borderWidths\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"InterleavedCirclesOptions\"},{\"kind\":128,\"name\":\"InterleavedShapes\",\"url\":\"classes/InterleavedShapes.html\",\"classes\":\"tsd-kind-class\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/InterleavedShapes.html#constructor\",\"classes\":\"tsd-kind-constructor tsd-parent-kind-class\",\"parent\":\"InterleavedShapes\"},{\"kind\":1024,\"name\":\"shape\",\"url\":\"classes/InterleavedShapes.html#shape\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"InterleavedShapes\"},{\"kind\":1024,\"name\":\"xys\",\"url\":\"classes/InterleavedShapes.html#xys\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"InterleavedShapes\"},{\"kind\":1024,\"name\":\"scales\",\"url\":\"classes/InterleavedShapes.html#scales\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"InterleavedShapes\"},{\"kind\":1024,\"name\":\"rotations\",\"url\":\"classes/InterleavedShapes.html#rotations\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"InterleavedShapes\"},{\"kind\":1024,\"name\":\"colors\",\"url\":\"classes/InterleavedShapes.html#colors\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"InterleavedShapes\"},{\"kind\":2048,\"name\":\"dispose\",\"url\":\"classes/InterleavedShapes.html#dispose\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"InterleavedShapes\"},{\"kind\":256,\"name\":\"InterleavedShapesOptions\",\"url\":\"interfaces/InterleavedShapesOptions.html\",\"classes\":\"tsd-kind-interface\"},{\"kind\":1024,\"name\":\"colors\",\"url\":\"interfaces/InterleavedShapesOptions.html#colors\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"InterleavedShapesOptions\"},{\"kind\":1024,\"name\":\"scales\",\"url\":\"interfaces/InterleavedShapesOptions.html#scales\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"InterleavedShapesOptions\"},{\"kind\":1024,\"name\":\"rotations\",\"url\":\"interfaces/InterleavedShapesOptions.html#rotations\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"InterleavedShapesOptions\"},{\"kind\":128,\"name\":\"LineSegments\",\"url\":\"classes/LineSegments.html\",\"classes\":\"tsd-kind-class\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/LineSegments.html#constructor\",\"classes\":\"tsd-kind-constructor tsd-parent-kind-class\",\"parent\":\"LineSegments\"},{\"kind\":1024,\"name\":\"points\",\"url\":\"classes/LineSegments.html#points\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"LineSegments\"},{\"kind\":1024,\"name\":\"widths\",\"url\":\"classes/LineSegments.html#widths\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"LineSegments\"},{\"kind\":1024,\"name\":\"colors\",\"url\":\"classes/LineSegments.html#colors\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"LineSegments\"},{\"kind\":1024,\"name\":\"segmentGeometry\",\"url\":\"classes/LineSegments.html#segmentGeometry\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"LineSegments\"},{\"kind\":2048,\"name\":\"dispose\",\"url\":\"classes/LineSegments.html#dispose\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"LineSegments\"},{\"kind\":256,\"name\":\"LineSegmentsOptions\",\"url\":\"interfaces/LineSegmentsOptions.html\",\"classes\":\"tsd-kind-interface\"},{\"kind\":1024,\"name\":\"widths\",\"url\":\"interfaces/LineSegmentsOptions.html#widths\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"LineSegmentsOptions\"},{\"kind\":1024,\"name\":\"colors\",\"url\":\"interfaces/LineSegmentsOptions.html#colors\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"LineSegmentsOptions\"},{\"kind\":128,\"name\":\"LineStrip\",\"url\":\"classes/LineStrip.html\",\"classes\":\"tsd-kind-class\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/LineStrip.html#constructor\",\"classes\":\"tsd-kind-constructor tsd-parent-kind-class\",\"parent\":\"LineStrip\"},{\"kind\":1024,\"name\":\"xs\",\"url\":\"classes/LineStrip.html#xs\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"LineStrip\"},{\"kind\":1024,\"name\":\"ys\",\"url\":\"classes/LineStrip.html#ys\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"LineStrip\"},{\"kind\":1024,\"name\":\"widths\",\"url\":\"classes/LineStrip.html#widths\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"LineStrip\"},{\"kind\":1024,\"name\":\"colors\",\"url\":\"classes/LineStrip.html#colors\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"LineStrip\"},{\"kind\":1024,\"name\":\"roundCapJoin\",\"url\":\"classes/LineStrip.html#roundCapJoin\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"LineStrip\"},{\"kind\":2048,\"name\":\"dispose\",\"url\":\"classes/LineStrip.html#dispose\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"LineStrip\"},{\"kind\":256,\"name\":\"LineStripOptions\",\"url\":\"interfaces/LineStripOptions.html\",\"classes\":\"tsd-kind-interface\"},{\"kind\":1024,\"name\":\"widths\",\"url\":\"interfaces/LineStripOptions.html#widths\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"LineStripOptions\"},{\"kind\":1024,\"name\":\"colors\",\"url\":\"interfaces/LineStripOptions.html#colors\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"LineStripOptions\"},{\"kind\":128,\"name\":\"TransparentLineStrip\",\"url\":\"classes/TransparentLineStrip.html\",\"classes\":\"tsd-kind-class\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/TransparentLineStrip.html#constructor\",\"classes\":\"tsd-kind-constructor tsd-parent-kind-class\",\"parent\":\"TransparentLineStrip\"},{\"kind\":1024,\"name\":\"xs\",\"url\":\"classes/TransparentLineStrip.html#xs\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"TransparentLineStrip\"},{\"kind\":1024,\"name\":\"ys\",\"url\":\"classes/TransparentLineStrip.html#ys\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"TransparentLineStrip\"},{\"kind\":1024,\"name\":\"width\",\"url\":\"classes/TransparentLineStrip.html#width\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"TransparentLineStrip\"},{\"kind\":1024,\"name\":\"color\",\"url\":\"classes/TransparentLineStrip.html#color\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"TransparentLineStrip\"},{\"kind\":1024,\"name\":\"segmentPositions\",\"url\":\"classes/TransparentLineStrip.html#segmentPositions\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"TransparentLineStrip\"},{\"kind\":1024,\"name\":\"segmentCells\",\"url\":\"classes/TransparentLineStrip.html#segmentCells\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"TransparentLineStrip\"},{\"kind\":1024,\"name\":\"joinIds\",\"url\":\"classes/TransparentLineStrip.html#joinIds\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"TransparentLineStrip\"},{\"kind\":1024,\"name\":\"joinCells\",\"url\":\"classes/TransparentLineStrip.html#joinCells\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"TransparentLineStrip\"},{\"kind\":1024,\"name\":\"capPositions\",\"url\":\"classes/TransparentLineStrip.html#capPositions\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"TransparentLineStrip\"},{\"kind\":1024,\"name\":\"capCells\",\"url\":\"classes/TransparentLineStrip.html#capCells\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"TransparentLineStrip\"},{\"kind\":2048,\"name\":\"dispose\",\"url\":\"classes/TransparentLineStrip.html#dispose\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"TransparentLineStrip\"},{\"kind\":256,\"name\":\"TransparentLineStripOptions\",\"url\":\"interfaces/TransparentLineStripOptions.html\",\"classes\":\"tsd-kind-interface\"},{\"kind\":1024,\"name\":\"width\",\"url\":\"interfaces/TransparentLineStripOptions.html#width\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"TransparentLineStripOptions\"},{\"kind\":1024,\"name\":\"color\",\"url\":\"interfaces/TransparentLineStripOptions.html#color\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"TransparentLineStripOptions\"},{\"kind\":128,\"name\":\"Rects\",\"url\":\"classes/Rects.html\",\"classes\":\"tsd-kind-class\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/Rects.html#constructor\",\"classes\":\"tsd-kind-constructor tsd-parent-kind-class\",\"parent\":\"Rects\"},{\"kind\":1024,\"name\":\"rects\",\"url\":\"classes/Rects.html#rects\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Rects\"},{\"kind\":1024,\"name\":\"colors\",\"url\":\"classes/Rects.html#colors\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Rects\"},{\"kind\":1024,\"name\":\"positionBuffer\",\"url\":\"classes/Rects.html#positionBuffer\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"Rects\"},{\"kind\":2048,\"name\":\"dispose\",\"url\":\"classes/Rects.html#dispose\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"Rects\"},{\"kind\":256,\"name\":\"RectsOptions\",\"url\":\"interfaces/RectsOptions.html\",\"classes\":\"tsd-kind-interface\"},{\"kind\":1024,\"name\":\"colors\",\"url\":\"interfaces/RectsOptions.html#colors\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"RectsOptions\"},{\"kind\":128,\"name\":\"Shapes\",\"url\":\"classes/Shapes.html\",\"classes\":\"tsd-kind-class\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/Shapes.html#constructor\",\"classes\":\"tsd-kind-constructor tsd-parent-kind-class\",\"parent\":\"Shapes\"},{\"kind\":1024,\"name\":\"shape\",\"url\":\"classes/Shapes.html#shape\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Shapes\"},{\"kind\":1024,\"name\":\"xs\",\"url\":\"classes/Shapes.html#xs\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Shapes\"},{\"kind\":1024,\"name\":\"ys\",\"url\":\"classes/Shapes.html#ys\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Shapes\"},{\"kind\":1024,\"name\":\"scales\",\"url\":\"classes/Shapes.html#scales\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Shapes\"},{\"kind\":1024,\"name\":\"rotations\",\"url\":\"classes/Shapes.html#rotations\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Shapes\"},{\"kind\":1024,\"name\":\"colors\",\"url\":\"classes/Shapes.html#colors\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Shapes\"},{\"kind\":2048,\"name\":\"dispose\",\"url\":\"classes/Shapes.html#dispose\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"Shapes\"},{\"kind\":256,\"name\":\"ShapesOptions\",\"url\":\"interfaces/ShapesOptions.html\",\"classes\":\"tsd-kind-interface\"},{\"kind\":1024,\"name\":\"colors\",\"url\":\"interfaces/ShapesOptions.html#colors\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"ShapesOptions\"},{\"kind\":1024,\"name\":\"scales\",\"url\":\"interfaces/ShapesOptions.html#scales\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"ShapesOptions\"},{\"kind\":1024,\"name\":\"rotations\",\"url\":\"interfaces/ShapesOptions.html#rotations\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"ShapesOptions\"},{\"kind\":128,\"name\":\"Text\",\"url\":\"classes/Text.html\",\"classes\":\"tsd-kind-class\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/Text.html#constructor\",\"classes\":\"tsd-kind-constructor tsd-parent-kind-class\",\"parent\":\"Text\"},{\"kind\":1024,\"name\":\"position\",\"url\":\"classes/Text.html#position\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Text\"},{\"kind\":1024,\"name\":\"size\",\"url\":\"classes/Text.html#size\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Text\"},{\"kind\":1024,\"name\":\"angle\",\"url\":\"classes/Text.html#angle\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Text\"},{\"kind\":1024,\"name\":\"color\",\"url\":\"classes/Text.html#color\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Text\"},{\"kind\":1024,\"name\":\"width\",\"url\":\"classes/Text.html#width\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Text\"},{\"kind\":1024,\"name\":\"height\",\"url\":\"classes/Text.html#height\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Text\"},{\"kind\":1024,\"name\":\"quad\",\"url\":\"classes/Text.html#quad\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"Text\"},{\"kind\":1024,\"name\":\"uv\",\"url\":\"classes/Text.html#uv\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"Text\"},{\"kind\":1024,\"name\":\"instances\",\"url\":\"classes/Text.html#instances\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"Text\"},{\"kind\":1024,\"name\":\"quadGeometry\",\"url\":\"classes/Text.html#quadGeometry\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"Text\"},{\"kind\":2048,\"name\":\"dispose\",\"url\":\"classes/Text.html#dispose\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"Text\"},{\"kind\":256,\"name\":\"TextOptions\",\"url\":\"interfaces/TextOptions.html\",\"classes\":\"tsd-kind-interface\"},{\"kind\":1024,\"name\":\"align\",\"url\":\"interfaces/TextOptions.html#align\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"TextOptions\"},{\"kind\":1024,\"name\":\"anchor\",\"url\":\"interfaces/TextOptions.html#anchor\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"TextOptions\"},{\"kind\":1024,\"name\":\"angle\",\"url\":\"interfaces/TextOptions.html#angle\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"TextOptions\"},{\"kind\":1024,\"name\":\"color\",\"url\":\"interfaces/TextOptions.html#color\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"TextOptions\"},{\"kind\":1024,\"name\":\"size\",\"url\":\"interfaces/TextOptions.html#size\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"TextOptions\"},{\"kind\":128,\"name\":\"Triangles\",\"url\":\"classes/Triangles.html\",\"classes\":\"tsd-kind-class\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/Triangles.html#constructor\",\"classes\":\"tsd-kind-constructor tsd-parent-kind-class\",\"parent\":\"Triangles\"},{\"kind\":1024,\"name\":\"vertices\",\"url\":\"classes/Triangles.html#vertices\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"Triangles\"},{\"kind\":1024,\"name\":\"color\",\"url\":\"classes/Triangles.html#color\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Triangles\"},{\"kind\":2048,\"name\":\"dispose\",\"url\":\"classes/Triangles.html#dispose\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"Triangles\"},{\"kind\":256,\"name\":\"TrianglesOptions\",\"url\":\"interfaces/TrianglesOptions.html\",\"classes\":\"tsd-kind-interface\"},{\"kind\":1024,\"name\":\"color\",\"url\":\"interfaces/TrianglesOptions.html#color\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"TrianglesOptions\"},{\"kind\":128,\"name\":\"VLines\",\"url\":\"classes/VLines.html\",\"classes\":\"tsd-kind-class\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/VLines.html#constructor\",\"classes\":\"tsd-kind-constructor tsd-parent-kind-class\",\"parent\":\"VLines\"},{\"kind\":1024,\"name\":\"lines\",\"url\":\"classes/VLines.html#lines\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"VLines\"},{\"kind\":1024,\"name\":\"widths\",\"url\":\"classes/VLines.html#widths\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"VLines\"},{\"kind\":1024,\"name\":\"colors\",\"url\":\"classes/VLines.html#colors\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"VLines\"},{\"kind\":1024,\"name\":\"segmentGeometry\",\"url\":\"classes/VLines.html#segmentGeometry\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"VLines\"},{\"kind\":2048,\"name\":\"dispose\",\"url\":\"classes/VLines.html#dispose\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"VLines\"},{\"kind\":256,\"name\":\"VLinesOptions\",\"url\":\"interfaces/VLinesOptions.html\",\"classes\":\"tsd-kind-interface\"},{\"kind\":1024,\"name\":\"widths\",\"url\":\"interfaces/VLinesOptions.html#widths\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"VLinesOptions\"},{\"kind\":1024,\"name\":\"colors\",\"url\":\"interfaces/VLinesOptions.html#colors\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"VLinesOptions\"},{\"kind\":128,\"name\":\"Wedges\",\"url\":\"classes/Wedges.html\",\"classes\":\"tsd-kind-class\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/Wedges.html#constructor\",\"classes\":\"tsd-kind-constructor tsd-parent-kind-class\",\"parent\":\"Wedges\"},{\"kind\":1024,\"name\":\"xys\",\"url\":\"classes/Wedges.html#xys\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Wedges\"},{\"kind\":1024,\"name\":\"angles\",\"url\":\"classes/Wedges.html#angles\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Wedges\"},{\"kind\":1024,\"name\":\"colors\",\"url\":\"classes/Wedges.html#colors\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Wedges\"},{\"kind\":1024,\"name\":\"radii\",\"url\":\"classes/Wedges.html#radii\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Wedges\"},{\"kind\":1024,\"name\":\"positionBuffer\",\"url\":\"classes/Wedges.html#positionBuffer\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"Wedges\"},{\"kind\":2048,\"name\":\"dispose\",\"url\":\"classes/Wedges.html#dispose\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"Wedges\"},{\"kind\":256,\"name\":\"WedgesOptions\",\"url\":\"interfaces/WedgesOptions.html\",\"classes\":\"tsd-kind-interface\"},{\"kind\":1024,\"name\":\"colors\",\"url\":\"interfaces/WedgesOptions.html#colors\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"WedgesOptions\"},{\"kind\":1024,\"name\":\"radii\",\"url\":\"interfaces/WedgesOptions.html#radii\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"WedgesOptions\"},{\"kind\":128,\"name\":\"Axis\",\"url\":\"classes/Axis.html\",\"classes\":\"tsd-kind-class\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/Axis.html#constructor\",\"classes\":\"tsd-kind-constructor tsd-parent-kind-class\",\"parent\":\"Axis\"},{\"kind\":1024,\"name\":\"axis\",\"url\":\"classes/Axis.html#axis\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"Axis\"},{\"kind\":1024,\"name\":\"texts\",\"url\":\"classes/Axis.html#texts\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"Axis\"},{\"kind\":1024,\"name\":\"ticks\",\"url\":\"classes/Axis.html#ticks\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"Axis\"},{\"kind\":1024,\"name\":\"minorTicks\",\"url\":\"classes/Axis.html#minorTicks\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"Axis\"},{\"kind\":2048,\"name\":\"dispose\",\"url\":\"classes/Axis.html#dispose\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Axis\"},{\"kind\":256,\"name\":\"AxisOptions\",\"url\":\"interfaces/AxisOptions.html\",\"classes\":\"tsd-kind-interface\"},{\"kind\":1024,\"name\":\"axisColor\",\"url\":\"interfaces/AxisOptions.html#axisColor\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"AxisOptions\"},{\"kind\":1024,\"name\":\"axisWidth\",\"url\":\"interfaces/AxisOptions.html#axisWidth\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"AxisOptions\"},{\"kind\":1024,\"name\":\"labelAnchor\",\"url\":\"interfaces/AxisOptions.html#labelAnchor\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"AxisOptions\"},{\"kind\":1024,\"name\":\"labelAngle\",\"url\":\"interfaces/AxisOptions.html#labelAngle\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"AxisOptions\"},{\"kind\":1024,\"name\":\"labelColor\",\"url\":\"interfaces/AxisOptions.html#labelColor\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"AxisOptions\"},{\"kind\":1024,\"name\":\"labelPad\",\"url\":\"interfaces/AxisOptions.html#labelPad\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"AxisOptions\"},{\"kind\":1024,\"name\":\"labelSide\",\"url\":\"interfaces/AxisOptions.html#labelSide\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"AxisOptions\"},{\"kind\":1024,\"name\":\"labelSize\",\"url\":\"interfaces/AxisOptions.html#labelSize\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"AxisOptions\"},{\"kind\":1024,\"name\":\"tickColor\",\"url\":\"interfaces/AxisOptions.html#tickColor\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"AxisOptions\"},{\"kind\":1024,\"name\":\"tickLength\",\"url\":\"interfaces/AxisOptions.html#tickLength\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"AxisOptions\"},{\"kind\":1024,\"name\":\"tickOffset\",\"url\":\"interfaces/AxisOptions.html#tickOffset\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"AxisOptions\"},{\"kind\":1024,\"name\":\"tickWidth\",\"url\":\"interfaces/AxisOptions.html#tickWidth\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"AxisOptions\"},{\"kind\":1024,\"name\":\"minorTicks\",\"url\":\"interfaces/AxisOptions.html#minorTicks\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"AxisOptions\"},{\"kind\":1024,\"name\":\"minorTickColor\",\"url\":\"interfaces/AxisOptions.html#minorTickColor\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"AxisOptions\"},{\"kind\":1024,\"name\":\"minorTickLength\",\"url\":\"interfaces/AxisOptions.html#minorTickLength\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"AxisOptions\"},{\"kind\":1024,\"name\":\"minorTickOffset\",\"url\":\"interfaces/AxisOptions.html#minorTickOffset\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"AxisOptions\"},{\"kind\":1024,\"name\":\"minorTickWidth\",\"url\":\"interfaces/AxisOptions.html#minorTickWidth\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"AxisOptions\"},{\"kind\":128,\"name\":\"Grid\",\"url\":\"classes/Grid.html\",\"classes\":\"tsd-kind-class\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/Grid.html#constructor\",\"classes\":\"tsd-kind-constructor tsd-parent-kind-class\",\"parent\":\"Grid\"},{\"kind\":1024,\"name\":\"grid\",\"url\":\"classes/Grid.html#grid\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"Grid\"},{\"kind\":2048,\"name\":\"dispose\",\"url\":\"classes/Grid.html#dispose\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Grid\"},{\"kind\":256,\"name\":\"GridOptions\",\"url\":\"interfaces/GridOptions.html\",\"classes\":\"tsd-kind-interface\"},{\"kind\":1024,\"name\":\"width\",\"url\":\"interfaces/GridOptions.html#width\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"GridOptions\"},{\"kind\":1024,\"name\":\"color\",\"url\":\"interfaces/GridOptions.html#color\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"GridOptions\"},{\"kind\":128,\"name\":\"OrthoAxis\",\"url\":\"classes/OrthoAxis.html\",\"classes\":\"tsd-kind-class\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/OrthoAxis.html#constructor\",\"classes\":\"tsd-kind-constructor tsd-parent-kind-class\",\"parent\":\"OrthoAxis\"},{\"kind\":1024,\"name\":\"info\",\"url\":\"classes/OrthoAxis.html#info\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"OrthoAxis\"},{\"kind\":1024,\"name\":\"axis\",\"url\":\"classes/OrthoAxis.html#axis\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"OrthoAxis\"},{\"kind\":2048,\"name\":\"dispose\",\"url\":\"classes/OrthoAxis.html#dispose\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"OrthoAxis\"},{\"kind\":256,\"name\":\"OrthoAxisOptions\",\"url\":\"interfaces/OrthoAxisOptions.html\",\"classes\":\"tsd-kind-interface\"},{\"kind\":1024,\"name\":\"axisHigh\",\"url\":\"interfaces/OrthoAxisOptions.html#axisHigh\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"OrthoAxisOptions\"},{\"kind\":1024,\"name\":\"axisIntercept\",\"url\":\"interfaces/OrthoAxisOptions.html#axisIntercept\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"OrthoAxisOptions\"},{\"kind\":1024,\"name\":\"axisLow\",\"url\":\"interfaces/OrthoAxisOptions.html#axisLow\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"OrthoAxisOptions\"},{\"kind\":1024,\"name\":\"labelFormatter\",\"url\":\"interfaces/OrthoAxisOptions.html#labelFormatter\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"OrthoAxisOptions\"},{\"kind\":65536,\"name\":\"__type\",\"url\":\"interfaces/OrthoAxisOptions.html#__type\",\"classes\":\"tsd-kind-type-literal tsd-parent-kind-interface\",\"parent\":\"OrthoAxisOptions\"},{\"kind\":1024,\"name\":\"minorTickCount\",\"url\":\"interfaces/OrthoAxisOptions.html#minorTickCount\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"OrthoAxisOptions\"},{\"kind\":1024,\"name\":\"tickOrigin\",\"url\":\"interfaces/OrthoAxisOptions.html#tickOrigin\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"OrthoAxisOptions\"},{\"kind\":1024,\"name\":\"tickStep\",\"url\":\"interfaces/OrthoAxisOptions.html#tickStep\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"OrthoAxisOptions\"},{\"kind\":1024,\"name\":\"axisColor\",\"url\":\"interfaces/OrthoAxisOptions.html#axisColor\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface tsd-is-inherited\",\"parent\":\"OrthoAxisOptions\"},{\"kind\":1024,\"name\":\"axisWidth\",\"url\":\"interfaces/OrthoAxisOptions.html#axisWidth\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface tsd-is-inherited\",\"parent\":\"OrthoAxisOptions\"},{\"kind\":1024,\"name\":\"labelAnchor\",\"url\":\"interfaces/OrthoAxisOptions.html#labelAnchor\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface tsd-is-inherited\",\"parent\":\"OrthoAxisOptions\"},{\"kind\":1024,\"name\":\"labelAngle\",\"url\":\"interfaces/OrthoAxisOptions.html#labelAngle\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface tsd-is-inherited\",\"parent\":\"OrthoAxisOptions\"},{\"kind\":1024,\"name\":\"labelColor\",\"url\":\"interfaces/OrthoAxisOptions.html#labelColor\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface tsd-is-inherited\",\"parent\":\"OrthoAxisOptions\"},{\"kind\":1024,\"name\":\"labelPad\",\"url\":\"interfaces/OrthoAxisOptions.html#labelPad\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface tsd-is-inherited\",\"parent\":\"OrthoAxisOptions\"},{\"kind\":1024,\"name\":\"labelSide\",\"url\":\"interfaces/OrthoAxisOptions.html#labelSide\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface tsd-is-inherited\",\"parent\":\"OrthoAxisOptions\"},{\"kind\":1024,\"name\":\"labelSize\",\"url\":\"interfaces/OrthoAxisOptions.html#labelSize\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface tsd-is-inherited\",\"parent\":\"OrthoAxisOptions\"},{\"kind\":1024,\"name\":\"tickColor\",\"url\":\"interfaces/OrthoAxisOptions.html#tickColor\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface tsd-is-inherited\",\"parent\":\"OrthoAxisOptions\"},{\"kind\":1024,\"name\":\"tickLength\",\"url\":\"interfaces/OrthoAxisOptions.html#tickLength\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface tsd-is-inherited\",\"parent\":\"OrthoAxisOptions\"},{\"kind\":1024,\"name\":\"tickOffset\",\"url\":\"interfaces/OrthoAxisOptions.html#tickOffset\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface tsd-is-inherited\",\"parent\":\"OrthoAxisOptions\"},{\"kind\":1024,\"name\":\"tickWidth\",\"url\":\"interfaces/OrthoAxisOptions.html#tickWidth\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface tsd-is-inherited\",\"parent\":\"OrthoAxisOptions\"},{\"kind\":1024,\"name\":\"minorTicks\",\"url\":\"interfaces/OrthoAxisOptions.html#minorTicks\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface tsd-is-inherited\",\"parent\":\"OrthoAxisOptions\"},{\"kind\":1024,\"name\":\"minorTickColor\",\"url\":\"interfaces/OrthoAxisOptions.html#minorTickColor\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface tsd-is-inherited\",\"parent\":\"OrthoAxisOptions\"},{\"kind\":1024,\"name\":\"minorTickLength\",\"url\":\"interfaces/OrthoAxisOptions.html#minorTickLength\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface tsd-is-inherited\",\"parent\":\"OrthoAxisOptions\"},{\"kind\":1024,\"name\":\"minorTickOffset\",\"url\":\"interfaces/OrthoAxisOptions.html#minorTickOffset\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface tsd-is-inherited\",\"parent\":\"OrthoAxisOptions\"},{\"kind\":1024,\"name\":\"minorTickWidth\",\"url\":\"interfaces/OrthoAxisOptions.html#minorTickWidth\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface tsd-is-inherited\",\"parent\":\"OrthoAxisOptions\"},{\"kind\":256,\"name\":\"OrthoAxisInfo\",\"url\":\"interfaces/OrthoAxisInfo.html\",\"classes\":\"tsd-kind-interface\"},{\"kind\":1024,\"name\":\"ticks\",\"url\":\"interfaces/OrthoAxisInfo.html#ticks\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"OrthoAxisInfo\"},{\"kind\":1024,\"name\":\"minorTicks\",\"url\":\"interfaces/OrthoAxisInfo.html#minorTicks\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"OrthoAxisInfo\"},{\"kind\":128,\"name\":\"Scissor\",\"url\":\"classes/Scissor.html\",\"classes\":\"tsd-kind-class\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/Scissor.html#constructor\",\"classes\":\"tsd-kind-constructor tsd-parent-kind-class\",\"parent\":\"Scissor\"},{\"kind\":1024,\"name\":\"_children\",\"url\":\"classes/Scissor.html#_children\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"Scissor\"},{\"kind\":1024,\"name\":\"scope\",\"url\":\"classes/Scissor.html#scope\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Scissor\"},{\"kind\":1024,\"name\":\"x\",\"url\":\"classes/Scissor.html#x\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Scissor\"},{\"kind\":1024,\"name\":\"y\",\"url\":\"classes/Scissor.html#y\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Scissor\"},{\"kind\":1024,\"name\":\"width\",\"url\":\"classes/Scissor.html#width\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Scissor\"},{\"kind\":1024,\"name\":\"height\",\"url\":\"classes/Scissor.html#height\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Scissor\"},{\"kind\":1024,\"name\":\"screenSpace\",\"url\":\"classes/Scissor.html#screenSpace\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Scissor\"},{\"kind\":2048,\"name\":\"dispose\",\"url\":\"classes/Scissor.html#dispose\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Scissor\"},{\"kind\":64,\"name\":\"createDefaultFont\",\"url\":\"functions/createDefaultFont.html\",\"classes\":\"tsd-kind-function\"},{\"kind\":128,\"name\":\"Font\",\"url\":\"classes/Font.html\",\"classes\":\"tsd-kind-class\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/Font.html#constructor\",\"classes\":\"tsd-kind-constructor tsd-parent-kind-class\",\"parent\":\"Font\"},{\"kind\":1024,\"name\":\"kernTable\",\"url\":\"classes/Font.html#kernTable\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"Font\"},{\"kind\":1024,\"name\":\"maxid\",\"url\":\"classes/Font.html#maxid\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"Font\"},{\"kind\":2048,\"name\":\"dispose\",\"url\":\"classes/Font.html#dispose\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"Font\"},{\"kind\":8388608,\"name\":\"default\",\"url\":\"index.html#default\",\"classes\":\"tsd-kind-reference\"}],\"index\":{\"version\":\"2.3.9\",\"fields\":[\"name\",\"comment\"],\"fieldVectors\":[[\"name/0\",[0,53.534]],[\"comment/0\",[]],[\"name/1\",[1,24.447]],[\"comment/1\",[]],[\"name/2\",[2,48.426]],[\"comment/2\",[]],[\"name/3\",[3,53.534]],[\"comment/3\",[]],[\"name/4\",[4,48.426]],[\"comment/4\",[]],[\"name/5\",[5,53.534]],[\"comment/5\",[]],[\"name/6\",[6,53.534]],[\"comment/6\",[]],[\"name/7\",[7,53.534]],[\"comment/7\",[]],[\"name/8\",[8,53.534]],[\"comment/8\",[]],[\"name/9\",[9,53.534]],[\"comment/9\",[]],[\"name/10\",[10,53.534]],[\"comment/10\",[]],[\"name/11\",[11,53.534]],[\"comment/11\",[]],[\"name/12\",[2,48.426]],[\"comment/12\",[]],[\"name/13\",[12,53.534]],[\"comment/13\",[]],[\"name/14\",[13,53.534]],[\"comment/14\",[]],[\"name/15\",[14,48.426]],[\"comment/15\",[]],[\"name/16\",[15,48.426]],[\"comment/16\",[]],[\"name/17\",[16,38.871]],[\"comment/17\",[]],[\"name/18\",[17,45.061]],[\"comment/18\",[]],[\"name/19\",[18,53.534]],[\"comment/19\",[]],[\"name/20\",[1,24.447]],[\"comment/20\",[]],[\"name/21\",[19,26.908]],[\"comment/21\",[]],[\"name/22\",[20,53.534]],[\"comment/22\",[]],[\"name/23\",[21,53.534]],[\"comment/23\",[]],[\"name/24\",[22,53.534]],[\"comment/24\",[]],[\"name/25\",[23,53.534]],[\"comment/25\",[]],[\"name/26\",[24,53.534]],[\"comment/26\",[]],[\"name/27\",[25,53.534]],[\"comment/27\",[]],[\"name/28\",[26,53.534]],[\"comment/28\",[]],[\"name/29\",[1,24.447]],[\"comment/29\",[]],[\"name/30\",[19,26.908]],[\"comment/30\",[]],[\"name/31\",[27,53.534]],[\"comment/31\",[]],[\"name/32\",[1,24.447]],[\"comment/32\",[]],[\"name/33\",[28,53.534]],[\"comment/33\",[]],[\"name/34\",[29,53.534]],[\"comment/34\",[]],[\"name/35\",[30,53.534]],[\"comment/35\",[]],[\"name/36\",[19,26.908]],[\"comment/36\",[]],[\"name/37\",[31,53.534]],[\"comment/37\",[]],[\"name/38\",[32,53.534]],[\"comment/38\",[]],[\"name/39\",[1,24.447]],[\"comment/39\",[]],[\"name/40\",[33,48.426]],[\"comment/40\",[]],[\"name/41\",[34,48.426]],[\"comment/41\",[]],[\"name/42\",[35,53.534]],[\"comment/42\",[]],[\"name/43\",[36,53.534]],[\"comment/43\",[]],[\"name/44\",[37,48.426]],[\"comment/44\",[]],[\"name/45\",[38,48.426]],[\"comment/45\",[]],[\"name/46\",[39,40.542]],[\"comment/46\",[]],[\"name/47\",[40,40.542]],[\"comment/47\",[]],[\"name/48\",[41,53.534]],[\"comment/48\",[]],[\"name/49\",[42,53.534]],[\"comment/49\",[]],[\"name/50\",[1,24.447]],[\"comment/50\",[]],[\"name/51\",[33,48.426]],[\"comment/51\",[]],[\"name/52\",[34,48.426]],[\"comment/52\",[]],[\"name/53\",[43,53.534]],[\"comment/53\",[]],[\"name/54\",[44,53.534]],[\"comment/54\",[]],[\"name/55\",[37,48.426]],[\"comment/55\",[]],[\"name/56\",[38,48.426]],[\"comment/56\",[]],[\"name/57\",[40,40.542]],[\"comment/57\",[]],[\"name/58\",[39,40.542]],[\"comment/58\",[]],[\"name/59\",[45,53.534]],[\"comment/59\",[]],[\"name/60\",[1,24.447]],[\"comment/60\",[]],[\"name/61\",[46,48.426]],[\"comment/61\",[]],[\"name/62\",[47,48.426]],[\"comment/62\",[]],[\"name/63\",[40,40.542]],[\"comment/63\",[]],[\"name/64\",[39,40.542]],[\"comment/64\",[]],[\"name/65\",[48,53.534]],[\"comment/65\",[]],[\"name/66\",[1,24.447]],[\"comment/66\",[]],[\"name/67\",[49,53.534]],[\"comment/67\",[]],[\"name/68\",[50,53.534]],[\"comment/68\",[]],[\"name/69\",[46,48.426]],[\"comment/69\",[]],[\"name/70\",[47,48.426]],[\"comment/70\",[]],[\"name/71\",[40,40.542]],[\"comment/71\",[]],[\"name/72\",[39,40.542]],[\"comment/72\",[]],[\"name/73\",[51,53.534]],[\"comment/73\",[]],[\"name/74\",[1,24.447]],[\"comment/74\",[]],[\"name/75\",[39,40.542]],[\"comment/75\",[]],[\"name/76\",[40,40.542]],[\"comment/76\",[]],[\"name/77\",[52,53.534]],[\"comment/77\",[]],[\"name/78\",[1,24.447]],[\"comment/78\",[]],[\"name/79\",[53,42.548]],[\"comment/79\",[]],[\"name/80\",[54,42.548]],[\"comment/80\",[]],[\"name/81\",[55,27.385]],[\"comment/81\",[]],[\"name/82\",[56,38.871]],[\"comment/82\",[]],[\"name/83\",[57,42.548]],[\"comment/83\",[]],[\"name/84\",[58,42.548]],[\"comment/84\",[]],[\"name/85\",[59,42.548]],[\"comment/85\",[]],[\"name/86\",[19,26.908]],[\"comment/86\",[]],[\"name/87\",[60,53.534]],[\"comment/87\",[]],[\"name/88\",[55,27.385]],[\"comment/88\",[]],[\"name/89\",[56,38.871]],[\"comment/89\",[]],[\"name/90\",[58,42.548]],[\"comment/90\",[]],[\"name/91\",[57,42.548]],[\"comment/91\",[]],[\"name/92\",[61,53.534]],[\"comment/92\",[]],[\"name/93\",[1,24.447]],[\"comment/93\",[]],[\"name/94\",[62,48.426]],[\"comment/94\",[]],[\"name/95\",[63,36.188]],[\"comment/95\",[]],[\"name/96\",[55,27.385]],[\"comment/96\",[]],[\"name/97\",[64,45.061]],[\"comment/97\",[]],[\"name/98\",[19,26.908]],[\"comment/98\",[]],[\"name/99\",[65,53.534]],[\"comment/99\",[]],[\"name/100\",[63,36.188]],[\"comment/100\",[]],[\"name/101\",[55,27.385]],[\"comment/101\",[]],[\"name/102\",[66,53.534]],[\"comment/102\",[]],[\"name/103\",[1,24.447]],[\"comment/103\",[]],[\"name/104\",[67,45.061]],[\"comment/104\",[]],[\"name/105\",[55,27.385]],[\"comment/105\",[]],[\"name/106\",[56,38.871]],[\"comment/106\",[]],[\"name/107\",[57,42.548]],[\"comment/107\",[]],[\"name/108\",[58,42.548]],[\"comment/108\",[]],[\"name/109\",[59,42.548]],[\"comment/109\",[]],[\"name/110\",[19,26.908]],[\"comment/110\",[]],[\"name/111\",[68,53.534]],[\"comment/111\",[]],[\"name/112\",[55,27.385]],[\"comment/112\",[]],[\"name/113\",[58,42.548]],[\"comment/113\",[]],[\"name/114\",[56,38.871]],[\"comment/114\",[]],[\"name/115\",[57,42.548]],[\"comment/115\",[]],[\"name/116\",[69,53.534]],[\"comment/116\",[]],[\"name/117\",[1,24.447]],[\"comment/117\",[]],[\"name/118\",[70,48.426]],[\"comment/118\",[]],[\"name/119\",[67,45.061]],[\"comment/119\",[]],[\"name/120\",[71,42.548]],[\"comment/120\",[]],[\"name/121\",[72,42.548]],[\"comment/121\",[]],[\"name/122\",[55,27.385]],[\"comment/122\",[]],[\"name/123\",[19,26.908]],[\"comment/123\",[]],[\"name/124\",[73,53.534]],[\"comment/124\",[]],[\"name/125\",[55,27.385]],[\"comment/125\",[]],[\"name/126\",[71,42.548]],[\"comment/126\",[]],[\"name/127\",[72,42.548]],[\"comment/127\",[]],[\"name/128\",[74,53.534]],[\"comment/128\",[]],[\"name/129\",[1,24.447]],[\"comment/129\",[]],[\"name/130\",[75,53.534]],[\"comment/130\",[]],[\"name/131\",[63,36.188]],[\"comment/131\",[]],[\"name/132\",[55,27.385]],[\"comment/132\",[]],[\"name/133\",[64,45.061]],[\"comment/133\",[]],[\"name/134\",[19,26.908]],[\"comment/134\",[]],[\"name/135\",[76,53.534]],[\"comment/135\",[]],[\"name/136\",[63,36.188]],[\"comment/136\",[]],[\"name/137\",[55,27.385]],[\"comment/137\",[]],[\"name/138\",[77,53.534]],[\"comment/138\",[]],[\"name/139\",[1,24.447]],[\"comment/139\",[]],[\"name/140\",[53,42.548]],[\"comment/140\",[]],[\"name/141\",[54,42.548]],[\"comment/141\",[]],[\"name/142\",[63,36.188]],[\"comment/142\",[]],[\"name/143\",[55,27.385]],[\"comment/143\",[]],[\"name/144\",[78,53.534]],[\"comment/144\",[]],[\"name/145\",[19,26.908]],[\"comment/145\",[]],[\"name/146\",[79,53.534]],[\"comment/146\",[]],[\"name/147\",[63,36.188]],[\"comment/147\",[]],[\"name/148\",[55,27.385]],[\"comment/148\",[]],[\"name/149\",[80,53.534]],[\"comment/149\",[]],[\"name/150\",[1,24.447]],[\"comment/150\",[]],[\"name/151\",[53,42.548]],[\"comment/151\",[]],[\"name/152\",[54,42.548]],[\"comment/152\",[]],[\"name/153\",[16,38.871]],[\"comment/153\",[]],[\"name/154\",[81,37.44]],[\"comment/154\",[]],[\"name/155\",[82,53.534]],[\"comment/155\",[]],[\"name/156\",[83,53.534]],[\"comment/156\",[]],[\"name/157\",[84,53.534]],[\"comment/157\",[]],[\"name/158\",[85,53.534]],[\"comment/158\",[]],[\"name/159\",[86,53.534]],[\"comment/159\",[]],[\"name/160\",[87,53.534]],[\"comment/160\",[]],[\"name/161\",[19,26.908]],[\"comment/161\",[]],[\"name/162\",[88,53.534]],[\"comment/162\",[]],[\"name/163\",[16,38.871]],[\"comment/163\",[]],[\"name/164\",[81,37.44]],[\"comment/164\",[]],[\"name/165\",[89,48.426]],[\"comment/165\",[]],[\"name/166\",[1,24.447]],[\"comment/166\",[]],[\"name/167\",[89,48.426]],[\"comment/167\",[]],[\"name/168\",[55,27.385]],[\"comment/168\",[]],[\"name/169\",[59,42.548]],[\"comment/169\",[]],[\"name/170\",[19,26.908]],[\"comment/170\",[]],[\"name/171\",[90,53.534]],[\"comment/171\",[]],[\"name/172\",[55,27.385]],[\"comment/172\",[]],[\"name/173\",[91,53.534]],[\"comment/173\",[]],[\"name/174\",[1,24.447]],[\"comment/174\",[]],[\"name/175\",[70,48.426]],[\"comment/175\",[]],[\"name/176\",[53,42.548]],[\"comment/176\",[]],[\"name/177\",[54,42.548]],[\"comment/177\",[]],[\"name/178\",[71,42.548]],[\"comment/178\",[]],[\"name/179\",[72,42.548]],[\"comment/179\",[]],[\"name/180\",[55,27.385]],[\"comment/180\",[]],[\"name/181\",[19,26.908]],[\"comment/181\",[]],[\"name/182\",[92,53.534]],[\"comment/182\",[]],[\"name/183\",[55,27.385]],[\"comment/183\",[]],[\"name/184\",[71,42.548]],[\"comment/184\",[]],[\"name/185\",[72,42.548]],[\"comment/185\",[]],[\"name/186\",[93,53.534]],[\"comment/186\",[]],[\"name/187\",[1,24.447]],[\"comment/187\",[]],[\"name/188\",[94,53.534]],[\"comment/188\",[]],[\"name/189\",[95,48.426]],[\"comment/189\",[]],[\"name/190\",[96,48.426]],[\"comment/190\",[]],[\"name/191\",[81,37.44]],[\"comment/191\",[]],[\"name/192\",[16,38.871]],[\"comment/192\",[]],[\"name/193\",[17,45.061]],[\"comment/193\",[]],[\"name/194\",[97,53.534]],[\"comment/194\",[]],[\"name/195\",[98,53.534]],[\"comment/195\",[]],[\"name/196\",[99,53.534]],[\"comment/196\",[]],[\"name/197\",[100,53.534]],[\"comment/197\",[]],[\"name/198\",[19,26.908]],[\"comment/198\",[]],[\"name/199\",[101,53.534]],[\"comment/199\",[]],[\"name/200\",[102,53.534]],[\"comment/200\",[]],[\"name/201\",[103,53.534]],[\"comment/201\",[]],[\"name/202\",[96,48.426]],[\"comment/202\",[]],[\"name/203\",[81,37.44]],[\"comment/203\",[]],[\"name/204\",[95,48.426]],[\"comment/204\",[]],[\"name/205\",[104,53.534]],[\"comment/205\",[]],[\"name/206\",[1,24.447]],[\"comment/206\",[]],[\"name/207\",[105,53.534]],[\"comment/207\",[]],[\"name/208\",[81,37.44]],[\"comment/208\",[]],[\"name/209\",[19,26.908]],[\"comment/209\",[]],[\"name/210\",[106,53.534]],[\"comment/210\",[]],[\"name/211\",[81,37.44]],[\"comment/211\",[]],[\"name/212\",[107,53.534]],[\"comment/212\",[]],[\"name/213\",[1,24.447]],[\"comment/213\",[]],[\"name/214\",[62,48.426]],[\"comment/214\",[]],[\"name/215\",[63,36.188]],[\"comment/215\",[]],[\"name/216\",[55,27.385]],[\"comment/216\",[]],[\"name/217\",[64,45.061]],[\"comment/217\",[]],[\"name/218\",[19,26.908]],[\"comment/218\",[]],[\"name/219\",[108,53.534]],[\"comment/219\",[]],[\"name/220\",[63,36.188]],[\"comment/220\",[]],[\"name/221\",[55,27.385]],[\"comment/221\",[]],[\"name/222\",[109,53.534]],[\"comment/222\",[]],[\"name/223\",[1,24.447]],[\"comment/223\",[]],[\"name/224\",[67,45.061]],[\"comment/224\",[]],[\"name/225\",[110,53.534]],[\"comment/225\",[]],[\"name/226\",[55,27.385]],[\"comment/226\",[]],[\"name/227\",[56,38.871]],[\"comment/227\",[]],[\"name/228\",[59,42.548]],[\"comment/228\",[]],[\"name/229\",[19,26.908]],[\"comment/229\",[]],[\"name/230\",[111,53.534]],[\"comment/230\",[]],[\"name/231\",[55,27.385]],[\"comment/231\",[]],[\"name/232\",[56,38.871]],[\"comment/232\",[]],[\"name/233\",[112,45.061]],[\"comment/233\",[]],[\"name/234\",[1,24.447]],[\"comment/234\",[]],[\"name/235\",[112,45.061]],[\"comment/235\",[]],[\"name/236\",[113,53.534]],[\"comment/236\",[]],[\"name/237\",[114,48.426]],[\"comment/237\",[]],[\"name/238\",[115,42.548]],[\"comment/238\",[]],[\"name/239\",[19,26.908]],[\"comment/239\",[]],[\"name/240\",[116,53.534]],[\"comment/240\",[]],[\"name/241\",[117,48.426]],[\"comment/241\",[]],[\"name/242\",[118,48.426]],[\"comment/242\",[]],[\"name/243\",[119,48.426]],[\"comment/243\",[]],[\"name/244\",[120,48.426]],[\"comment/244\",[]],[\"name/245\",[121,48.426]],[\"comment/245\",[]],[\"name/246\",[122,48.426]],[\"comment/246\",[]],[\"name/247\",[123,48.426]],[\"comment/247\",[]],[\"name/248\",[124,48.426]],[\"comment/248\",[]],[\"name/249\",[125,48.426]],[\"comment/249\",[]],[\"name/250\",[126,48.426]],[\"comment/250\",[]],[\"name/251\",[127,48.426]],[\"comment/251\",[]],[\"name/252\",[128,48.426]],[\"comment/252\",[]],[\"name/253\",[115,42.548]],[\"comment/253\",[]],[\"name/254\",[129,48.426]],[\"comment/254\",[]],[\"name/255\",[130,48.426]],[\"comment/255\",[]],[\"name/256\",[131,48.426]],[\"comment/256\",[]],[\"name/257\",[132,48.426]],[\"comment/257\",[]],[\"name/258\",[133,48.426]],[\"comment/258\",[]],[\"name/259\",[1,24.447]],[\"comment/259\",[]],[\"name/260\",[133,48.426]],[\"comment/260\",[]],[\"name/261\",[19,26.908]],[\"comment/261\",[]],[\"name/262\",[134,53.534]],[\"comment/262\",[]],[\"name/263\",[16,38.871]],[\"comment/263\",[]],[\"name/264\",[81,37.44]],[\"comment/264\",[]],[\"name/265\",[135,53.534]],[\"comment/265\",[]],[\"name/266\",[1,24.447]],[\"comment/266\",[]],[\"name/267\",[136,53.534]],[\"comment/267\",[]],[\"name/268\",[112,45.061]],[\"comment/268\",[]],[\"name/269\",[19,26.908]],[\"comment/269\",[]],[\"name/270\",[137,53.534]],[\"comment/270\",[]],[\"name/271\",[138,53.534]],[\"comment/271\",[]],[\"name/272\",[139,53.534]],[\"comment/272\",[]],[\"name/273\",[140,53.534]],[\"comment/273\",[]],[\"name/274\",[141,53.534]],[\"comment/274\",[]],[\"name/275\",[142,53.534]],[\"comment/275\",[]],[\"name/276\",[143,53.534]],[\"comment/276\",[]],[\"name/277\",[144,53.534]],[\"comment/277\",[]],[\"name/278\",[145,53.534]],[\"comment/278\",[]],[\"name/279\",[117,48.426]],[\"comment/279\",[]],[\"name/280\",[118,48.426]],[\"comment/280\",[]],[\"name/281\",[119,48.426]],[\"comment/281\",[]],[\"name/282\",[120,48.426]],[\"comment/282\",[]],[\"name/283\",[121,48.426]],[\"comment/283\",[]],[\"name/284\",[122,48.426]],[\"comment/284\",[]],[\"name/285\",[123,48.426]],[\"comment/285\",[]],[\"name/286\",[124,48.426]],[\"comment/286\",[]],[\"name/287\",[125,48.426]],[\"comment/287\",[]],[\"name/288\",[126,48.426]],[\"comment/288\",[]],[\"name/289\",[127,48.426]],[\"comment/289\",[]],[\"name/290\",[128,48.426]],[\"comment/290\",[]],[\"name/291\",[115,42.548]],[\"comment/291\",[]],[\"name/292\",[129,48.426]],[\"comment/292\",[]],[\"name/293\",[130,48.426]],[\"comment/293\",[]],[\"name/294\",[131,48.426]],[\"comment/294\",[]],[\"name/295\",[132,48.426]],[\"comment/295\",[]],[\"name/296\",[146,53.534]],[\"comment/296\",[]],[\"name/297\",[114,48.426]],[\"comment/297\",[]],[\"name/298\",[115,42.548]],[\"comment/298\",[]],[\"name/299\",[147,53.534]],[\"comment/299\",[]],[\"name/300\",[1,24.447]],[\"comment/300\",[]],[\"name/301\",[148,53.534]],[\"comment/301\",[]],[\"name/302\",[4,48.426]],[\"comment/302\",[]],[\"name/303\",[14,48.426]],[\"comment/303\",[]],[\"name/304\",[15,48.426]],[\"comment/304\",[]],[\"name/305\",[16,38.871]],[\"comment/305\",[]],[\"name/306\",[17,45.061]],[\"comment/306\",[]],[\"name/307\",[149,53.534]],[\"comment/307\",[]],[\"name/308\",[19,26.908]],[\"comment/308\",[]],[\"name/309\",[150,53.534]],[\"comment/309\",[]],[\"name/310\",[151,53.534]],[\"comment/310\",[]],[\"name/311\",[1,24.447]],[\"comment/311\",[]],[\"name/312\",[152,53.534]],[\"comment/312\",[]],[\"name/313\",[153,53.534]],[\"comment/313\",[]],[\"name/314\",[19,26.908]],[\"comment/314\",[]],[\"name/315\",[154,53.534]],[\"comment/315\",[]]],\"invertedIndex\":[[\"__type\",{\"_index\":142,\"name\":{\"275\":{}},\"comment\":{}}],[\"_children\",{\"_index\":148,\"name\":{\"301\":{}},\"comment\":{}}],[\"_domain\",{\"_index\":33,\"name\":{\"40\":{},\"51\":{}},\"comment\":{}}],[\"_range\",{\"_index\":34,\"name\":{\"41\":{},\"52\":{}},\"comment\":{}}],[\"align\",{\"_index\":102,\"name\":{\"200\":{}},\"comment\":{}}],[\"alpha\",{\"_index\":12,\"name\":{\"13\":{}},\"comment\":{}}],[\"anchor\",{\"_index\":103,\"name\":{\"201\":{}},\"comment\":{}}],[\"angle\",{\"_index\":96,\"name\":{\"190\":{},\"202\":{}},\"comment\":{}}],[\"angles\",{\"_index\":110,\"name\":{\"225\":{}},\"comment\":{}}],[\"angularscale\",{\"_index\":50,\"name\":{\"68\":{}},\"comment\":{}}],[\"assertfits\",{\"_index\":31,\"name\":{\"37\":{}},\"comment\":{}}],[\"axis\",{\"_index\":112,\"name\":{\"233\":{},\"235\":{},\"268\":{}},\"comment\":{}}],[\"axiscolor\",{\"_index\":117,\"name\":{\"241\":{},\"279\":{}},\"comment\":{}}],[\"axishigh\",{\"_index\":138,\"name\":{\"271\":{}},\"comment\":{}}],[\"axisintercept\",{\"_index\":139,\"name\":{\"272\":{}},\"comment\":{}}],[\"axislow\",{\"_index\":140,\"name\":{\"273\":{}},\"comment\":{}}],[\"axisoptions\",{\"_index\":116,\"name\":{\"240\":{}},\"comment\":{}}],[\"axiswidth\",{\"_index\":118,\"name\":{\"242\":{},\"280\":{}},\"comment\":{}}],[\"base\",{\"_index\":44,\"name\":{\"54\":{}},\"comment\":{}}],[\"bordercolors\",{\"_index\":58,\"name\":{\"84\":{},\"90\":{},\"108\":{},\"113\":{}},\"comment\":{}}],[\"borderwidths\",{\"_index\":57,\"name\":{\"83\":{},\"91\":{},\"107\":{},\"115\":{}},\"comment\":{}}],[\"candygraph\",{\"_index\":0,\"name\":{\"0\":{}},\"comment\":{}}],[\"candygraphoptions\",{\"_index\":11,\"name\":{\"11\":{}},\"comment\":{}}],[\"canvas\",{\"_index\":2,\"name\":{\"2\":{},\"12\":{}},\"comment\":{}}],[\"capcells\",{\"_index\":87,\"name\":{\"160\":{}},\"comment\":{}}],[\"cappositions\",{\"_index\":86,\"name\":{\"159\":{}},\"comment\":{}}],[\"cartesiancoordinatesystem\",{\"_index\":45,\"name\":{\"59\":{}},\"comment\":{}}],[\"circles\",{\"_index\":52,\"name\":{\"77\":{}},\"comment\":{}}],[\"circlesoptions\",{\"_index\":60,\"name\":{\"87\":{}},\"comment\":{}}],[\"clear\",{\"_index\":5,\"name\":{\"5\":{}},\"comment\":{}}],[\"color\",{\"_index\":81,\"name\":{\"154\":{},\"164\":{},\"191\":{},\"203\":{},\"208\":{},\"211\":{},\"264\":{}},\"comment\":{}}],[\"colors\",{\"_index\":55,\"name\":{\"81\":{},\"88\":{},\"96\":{},\"101\":{},\"105\":{},\"112\":{},\"122\":{},\"125\":{},\"132\":{},\"137\":{},\"143\":{},\"148\":{},\"168\":{},\"172\":{},\"180\":{},\"183\":{},\"216\":{},\"221\":{},\"226\":{},\"231\":{}},\"comment\":{}}],[\"commandsbycoords\",{\"_index\":3,\"name\":{\"3\":{}},\"comment\":{}}],[\"composite\",{\"_index\":18,\"name\":{\"19\":{}},\"comment\":{}}],[\"constructor\",{\"_index\":1,\"name\":{\"1\":{},\"20\":{},\"29\":{},\"32\":{},\"39\":{},\"50\":{},\"60\":{},\"66\":{},\"74\":{},\"78\":{},\"93\":{},\"103\":{},\"117\":{},\"129\":{},\"139\":{},\"150\":{},\"166\":{},\"174\":{},\"187\":{},\"206\":{},\"213\":{},\"223\":{},\"234\":{},\"259\":{},\"266\":{},\"300\":{},\"311\":{}},\"comment\":{}}],[\"conversion\",{\"_index\":43,\"name\":{\"53\":{}},\"comment\":{}}],[\"coordinatesystem\",{\"_index\":51,\"name\":{\"73\":{}},\"comment\":{}}],[\"copyto\",{\"_index\":8,\"name\":{\"8\":{}},\"comment\":{}}],[\"createdefaultfont\",{\"_index\":150,\"name\":{\"309\":{}},\"comment\":{}}],[\"dataset\",{\"_index\":27,\"name\":{\"31\":{}},\"comment\":{}}],[\"default\",{\"_index\":154,\"name\":{\"315\":{}},\"comment\":{}}],[\"destroy\",{\"_index\":7,\"name\":{\"7\":{}},\"comment\":{}}],[\"dispose\",{\"_index\":19,\"name\":{\"21\":{},\"30\":{},\"36\":{},\"86\":{},\"98\":{},\"110\":{},\"123\":{},\"134\":{},\"145\":{},\"161\":{},\"170\":{},\"181\":{},\"198\":{},\"209\":{},\"218\":{},\"229\":{},\"239\":{},\"261\":{},\"269\":{},\"308\":{},\"314\":{}},\"comment\":{}}],[\"disposed\",{\"_index\":29,\"name\":{\"34\":{}},\"comment\":{}}],[\"domain\",{\"_index\":37,\"name\":{\"44\":{},\"55\":{}},\"comment\":{}}],[\"font\",{\"_index\":151,\"name\":{\"310\":{}},\"comment\":{}}],[\"getnamedcommands\",{\"_index\":9,\"name\":{\"9\":{}},\"comment\":{}}],[\"grid\",{\"_index\":133,\"name\":{\"258\":{},\"260\":{}},\"comment\":{}}],[\"gridoptions\",{\"_index\":134,\"name\":{\"262\":{}},\"comment\":{}}],[\"height\",{\"_index\":17,\"name\":{\"18\":{},\"193\":{},\"306\":{}},\"comment\":{}}],[\"hlines\",{\"_index\":61,\"name\":{\"92\":{}},\"comment\":{}}],[\"hlinesoptions\",{\"_index\":65,\"name\":{\"99\":{}},\"comment\":{}}],[\"info\",{\"_index\":136,\"name\":{\"267\":{}},\"comment\":{}}],[\"instances\",{\"_index\":99,\"name\":{\"196\":{}},\"comment\":{}}],[\"interleavedcircles\",{\"_index\":66,\"name\":{\"102\":{}},\"comment\":{}}],[\"interleavedcirclesoptions\",{\"_index\":68,\"name\":{\"111\":{}},\"comment\":{}}],[\"interleavedshapes\",{\"_index\":69,\"name\":{\"116\":{}},\"comment\":{}}],[\"interleavedshapesoptions\",{\"_index\":73,\"name\":{\"124\":{}},\"comment\":{}}],[\"joincells\",{\"_index\":85,\"name\":{\"158\":{}},\"comment\":{}}],[\"joinids\",{\"_index\":84,\"name\":{\"157\":{}},\"comment\":{}}],[\"kerntable\",{\"_index\":152,\"name\":{\"312\":{}},\"comment\":{}}],[\"labelanchor\",{\"_index\":119,\"name\":{\"243\":{},\"281\":{}},\"comment\":{}}],[\"labelangle\",{\"_index\":120,\"name\":{\"244\":{},\"282\":{}},\"comment\":{}}],[\"labelcolor\",{\"_index\":121,\"name\":{\"245\":{},\"283\":{}},\"comment\":{}}],[\"labelformatter\",{\"_index\":141,\"name\":{\"274\":{}},\"comment\":{}}],[\"labelpad\",{\"_index\":122,\"name\":{\"246\":{},\"284\":{}},\"comment\":{}}],[\"labelside\",{\"_index\":123,\"name\":{\"247\":{},\"285\":{}},\"comment\":{}}],[\"labelsize\",{\"_index\":124,\"name\":{\"248\":{},\"286\":{}},\"comment\":{}}],[\"length\",{\"_index\":28,\"name\":{\"33\":{}},\"comment\":{}}],[\"linearscale\",{\"_index\":32,\"name\":{\"38\":{}},\"comment\":{}}],[\"lines\",{\"_index\":62,\"name\":{\"94\":{},\"214\":{}},\"comment\":{}}],[\"linesegments\",{\"_index\":74,\"name\":{\"128\":{}},\"comment\":{}}],[\"linesegmentsoptions\",{\"_index\":76,\"name\":{\"135\":{}},\"comment\":{}}],[\"linestrip\",{\"_index\":77,\"name\":{\"138\":{}},\"comment\":{}}],[\"linestripoptions\",{\"_index\":79,\"name\":{\"146\":{}},\"comment\":{}}],[\"logscale\",{\"_index\":42,\"name\":{\"49\":{}},\"comment\":{}}],[\"maxid\",{\"_index\":153,\"name\":{\"313\":{}},\"comment\":{}}],[\"minortickcolor\",{\"_index\":129,\"name\":{\"254\":{},\"292\":{}},\"comment\":{}}],[\"minortickcount\",{\"_index\":143,\"name\":{\"276\":{}},\"comment\":{}}],[\"minorticklength\",{\"_index\":130,\"name\":{\"255\":{},\"293\":{}},\"comment\":{}}],[\"minortickoffset\",{\"_index\":131,\"name\":{\"256\":{},\"294\":{}},\"comment\":{}}],[\"minorticks\",{\"_index\":115,\"name\":{\"238\":{},\"253\":{},\"291\":{},\"298\":{}},\"comment\":{}}],[\"minortickwidth\",{\"_index\":132,\"name\":{\"257\":{},\"295\":{}},\"comment\":{}}],[\"nameddrawcommands\",{\"_index\":25,\"name\":{\"27\":{}},\"comment\":{}}],[\"numberarray\",{\"_index\":21,\"name\":{\"23\":{}},\"comment\":{}}],[\"orthoaxis\",{\"_index\":135,\"name\":{\"265\":{}},\"comment\":{}}],[\"orthoaxisinfo\",{\"_index\":146,\"name\":{\"296\":{}},\"comment\":{}}],[\"orthoaxisoptions\",{\"_index\":137,\"name\":{\"270\":{}},\"comment\":{}}],[\"points\",{\"_index\":75,\"name\":{\"130\":{}},\"comment\":{}}],[\"polarcoordinatesystem\",{\"_index\":48,\"name\":{\"65\":{}},\"comment\":{}}],[\"position\",{\"_index\":94,\"name\":{\"188\":{}},\"comment\":{}}],[\"positionbuffer\",{\"_index\":59,\"name\":{\"85\":{},\"109\":{},\"169\":{},\"228\":{}},\"comment\":{}}],[\"primitive\",{\"_index\":26,\"name\":{\"28\":{}},\"comment\":{}}],[\"qdomain\",{\"_index\":35,\"name\":{\"42\":{}},\"comment\":{}}],[\"qrange\",{\"_index\":36,\"name\":{\"43\":{}},\"comment\":{}}],[\"quad\",{\"_index\":97,\"name\":{\"194\":{}},\"comment\":{}}],[\"quadgeometry\",{\"_index\":100,\"name\":{\"197\":{}},\"comment\":{}}],[\"radialscale\",{\"_index\":49,\"name\":{\"67\":{}},\"comment\":{}}],[\"radii\",{\"_index\":56,\"name\":{\"82\":{},\"89\":{},\"106\":{},\"114\":{},\"227\":{},\"232\":{}},\"comment\":{}}],[\"range\",{\"_index\":38,\"name\":{\"45\":{},\"56\":{}},\"comment\":{}}],[\"rects\",{\"_index\":89,\"name\":{\"165\":{},\"167\":{}},\"comment\":{}}],[\"rectsoptions\",{\"_index\":90,\"name\":{\"171\":{}},\"comment\":{}}],[\"recursiverender\",{\"_index\":10,\"name\":{\"10\":{}},\"comment\":{}}],[\"render\",{\"_index\":6,\"name\":{\"6\":{}},\"comment\":{}}],[\"renderable\",{\"_index\":20,\"name\":{\"22\":{}},\"comment\":{}}],[\"rotations\",{\"_index\":72,\"name\":{\"121\":{},\"127\":{},\"179\":{},\"185\":{}},\"comment\":{}}],[\"roundcapjoin\",{\"_index\":78,\"name\":{\"144\":{}},\"comment\":{}}],[\"scales\",{\"_index\":71,\"name\":{\"120\":{},\"126\":{},\"178\":{},\"184\":{}},\"comment\":{}}],[\"scissor\",{\"_index\":147,\"name\":{\"299\":{}},\"comment\":{}}],[\"scope\",{\"_index\":4,\"name\":{\"4\":{},\"302\":{}},\"comment\":{}}],[\"screenspace\",{\"_index\":149,\"name\":{\"307\":{}},\"comment\":{}}],[\"segmentcells\",{\"_index\":83,\"name\":{\"156\":{}},\"comment\":{}}],[\"segmentgeometry\",{\"_index\":64,\"name\":{\"97\":{},\"133\":{},\"217\":{}},\"comment\":{}}],[\"segmentpositions\",{\"_index\":82,\"name\":{\"155\":{}},\"comment\":{}}],[\"shape\",{\"_index\":70,\"name\":{\"118\":{},\"175\":{}},\"comment\":{}}],[\"shapes\",{\"_index\":91,\"name\":{\"173\":{}},\"comment\":{}}],[\"shapesoptions\",{\"_index\":92,\"name\":{\"182\":{}},\"comment\":{}}],[\"size\",{\"_index\":95,\"name\":{\"189\":{},\"204\":{}},\"comment\":{}}],[\"text\",{\"_index\":93,\"name\":{\"186\":{}},\"comment\":{}}],[\"textoptions\",{\"_index\":101,\"name\":{\"199\":{}},\"comment\":{}}],[\"texts\",{\"_index\":113,\"name\":{\"236\":{}},\"comment\":{}}],[\"tickcolor\",{\"_index\":125,\"name\":{\"249\":{},\"287\":{}},\"comment\":{}}],[\"ticklength\",{\"_index\":126,\"name\":{\"250\":{},\"288\":{}},\"comment\":{}}],[\"tickoffset\",{\"_index\":127,\"name\":{\"251\":{},\"289\":{}},\"comment\":{}}],[\"tickorigin\",{\"_index\":144,\"name\":{\"277\":{}},\"comment\":{}}],[\"ticks\",{\"_index\":114,\"name\":{\"237\":{},\"297\":{}},\"comment\":{}}],[\"tickstep\",{\"_index\":145,\"name\":{\"278\":{}},\"comment\":{}}],[\"tickwidth\",{\"_index\":128,\"name\":{\"252\":{},\"290\":{}},\"comment\":{}}],[\"todomain\",{\"_index\":39,\"name\":{\"46\":{},\"58\":{},\"64\":{},\"72\":{},\"75\":{}},\"comment\":{}}],[\"torange\",{\"_index\":40,\"name\":{\"47\":{},\"57\":{},\"63\":{},\"71\":{},\"76\":{}},\"comment\":{}}],[\"transparentlinestrip\",{\"_index\":80,\"name\":{\"149\":{}},\"comment\":{}}],[\"transparentlinestripoptions\",{\"_index\":88,\"name\":{\"162\":{}},\"comment\":{}}],[\"triangles\",{\"_index\":104,\"name\":{\"205\":{}},\"comment\":{}}],[\"trianglesoptions\",{\"_index\":106,\"name\":{\"210\":{}},\"comment\":{}}],[\"update\",{\"_index\":30,\"name\":{\"35\":{}},\"comment\":{}}],[\"updateq\",{\"_index\":41,\"name\":{\"48\":{}},\"comment\":{}}],[\"uv\",{\"_index\":98,\"name\":{\"195\":{}},\"comment\":{}}],[\"vector2\",{\"_index\":22,\"name\":{\"24\":{}},\"comment\":{}}],[\"vector3\",{\"_index\":23,\"name\":{\"25\":{}},\"comment\":{}}],[\"vector4\",{\"_index\":24,\"name\":{\"26\":{}},\"comment\":{}}],[\"vertices\",{\"_index\":105,\"name\":{\"207\":{}},\"comment\":{}}],[\"viewport\",{\"_index\":13,\"name\":{\"14\":{}},\"comment\":{}}],[\"vlines\",{\"_index\":107,\"name\":{\"212\":{}},\"comment\":{}}],[\"vlinesoptions\",{\"_index\":108,\"name\":{\"219\":{}},\"comment\":{}}],[\"wedges\",{\"_index\":109,\"name\":{\"222\":{}},\"comment\":{}}],[\"wedgesoptions\",{\"_index\":111,\"name\":{\"230\":{}},\"comment\":{}}],[\"width\",{\"_index\":16,\"name\":{\"17\":{},\"153\":{},\"163\":{},\"192\":{},\"263\":{},\"305\":{}},\"comment\":{}}],[\"widths\",{\"_index\":63,\"name\":{\"95\":{},\"100\":{},\"131\":{},\"136\":{},\"142\":{},\"147\":{},\"215\":{},\"220\":{}},\"comment\":{}}],[\"x\",{\"_index\":14,\"name\":{\"15\":{},\"303\":{}},\"comment\":{}}],[\"xs\",{\"_index\":53,\"name\":{\"79\":{},\"140\":{},\"151\":{},\"176\":{}},\"comment\":{}}],[\"xscale\",{\"_index\":46,\"name\":{\"61\":{},\"69\":{}},\"comment\":{}}],[\"xys\",{\"_index\":67,\"name\":{\"104\":{},\"119\":{},\"224\":{}},\"comment\":{}}],[\"y\",{\"_index\":15,\"name\":{\"16\":{},\"304\":{}},\"comment\":{}}],[\"ys\",{\"_index\":54,\"name\":{\"80\":{},\"141\":{},\"152\":{},\"177\":{}},\"comment\":{}}],[\"yscale\",{\"_index\":47,\"name\":{\"62\":{},\"70\":{}},\"comment\":{}}]],\"pipeline\":[]}}");
+window.searchData = JSON.parse("{\"kinds\":{\"64\":\"Function\",\"128\":\"Class\",\"256\":\"Interface\",\"512\":\"Constructor\",\"1024\":\"Property\",\"2048\":\"Method\",\"65536\":\"Type literal\",\"262144\":\"Accessor\",\"4194304\":\"Type alias\",\"8388608\":\"Reference\"},\"rows\":[{\"kind\":128,\"name\":\"CandyGraph\",\"url\":\"classes/CandyGraph.html\",\"classes\":\"tsd-kind-class\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/CandyGraph.html#constructor\",\"classes\":\"tsd-kind-constructor tsd-parent-kind-class\",\"parent\":\"CandyGraph\"},{\"kind\":1024,\"name\":\"canvas\",\"url\":\"classes/CandyGraph.html#canvas\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"CandyGraph\"},{\"kind\":1024,\"name\":\"commandsByCoords\",\"url\":\"classes/CandyGraph.html#commandsByCoords\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"CandyGraph\"},{\"kind\":1024,\"name\":\"scope\",\"url\":\"classes/CandyGraph.html#scope\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"CandyGraph\"},{\"kind\":2048,\"name\":\"clear\",\"url\":\"classes/CandyGraph.html#clear\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"CandyGraph\"},{\"kind\":2048,\"name\":\"render\",\"url\":\"classes/CandyGraph.html#render\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"CandyGraph\"},{\"kind\":2048,\"name\":\"destroy\",\"url\":\"classes/CandyGraph.html#destroy\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"CandyGraph\"},{\"kind\":2048,\"name\":\"copyTo\",\"url\":\"classes/CandyGraph.html#copyTo\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"CandyGraph\"},{\"kind\":2048,\"name\":\"getNamedCommands\",\"url\":\"classes/CandyGraph.html#getNamedCommands\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-private\",\"parent\":\"CandyGraph\"},{\"kind\":2048,\"name\":\"recursiveRender\",\"url\":\"classes/CandyGraph.html#recursiveRender\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-private\",\"parent\":\"CandyGraph\"},{\"kind\":256,\"name\":\"CandyGraphOptions\",\"url\":\"interfaces/CandyGraphOptions.html\",\"classes\":\"tsd-kind-interface\"},{\"kind\":1024,\"name\":\"canvas\",\"url\":\"interfaces/CandyGraphOptions.html#canvas\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"CandyGraphOptions\"},{\"kind\":1024,\"name\":\"alpha\",\"url\":\"interfaces/CandyGraphOptions.html#alpha\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"CandyGraphOptions\"},{\"kind\":256,\"name\":\"Viewport\",\"url\":\"interfaces/Viewport.html\",\"classes\":\"tsd-kind-interface\"},{\"kind\":1024,\"name\":\"x\",\"url\":\"interfaces/Viewport.html#x\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"Viewport\"},{\"kind\":1024,\"name\":\"y\",\"url\":\"interfaces/Viewport.html#y\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"Viewport\"},{\"kind\":1024,\"name\":\"width\",\"url\":\"interfaces/Viewport.html#width\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"Viewport\"},{\"kind\":1024,\"name\":\"height\",\"url\":\"interfaces/Viewport.html#height\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"Viewport\"},{\"kind\":4194304,\"name\":\"Renderable\",\"url\":\"types/Renderable.html\",\"classes\":\"tsd-kind-type-alias\"},{\"kind\":4194304,\"name\":\"NumberArray\",\"url\":\"types/NumberArray.html\",\"classes\":\"tsd-kind-type-alias\"},{\"kind\":4194304,\"name\":\"Vector2\",\"url\":\"types/Vector2.html\",\"classes\":\"tsd-kind-type-alias\"},{\"kind\":4194304,\"name\":\"Vector3\",\"url\":\"types/Vector3.html\",\"classes\":\"tsd-kind-type-alias\"},{\"kind\":4194304,\"name\":\"Vector4\",\"url\":\"types/Vector4.html\",\"classes\":\"tsd-kind-type-alias\"},{\"kind\":4194304,\"name\":\"NamedDrawCommands\",\"url\":\"types/NamedDrawCommands.html\",\"classes\":\"tsd-kind-type-alias\"},{\"kind\":128,\"name\":\"Primitive\",\"url\":\"classes/Primitive.html\",\"classes\":\"tsd-kind-class\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/Primitive.html#constructor\",\"classes\":\"tsd-kind-constructor tsd-parent-kind-class\",\"parent\":\"Primitive\"},{\"kind\":2048,\"name\":\"dispose\",\"url\":\"classes/Primitive.html#dispose\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"Primitive\"},{\"kind\":128,\"name\":\"Composite\",\"url\":\"classes/Composite.html\",\"classes\":\"tsd-kind-class\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/Composite.html#constructor\",\"classes\":\"tsd-kind-constructor tsd-parent-kind-class\",\"parent\":\"Composite\"},{\"kind\":2048,\"name\":\"dispose\",\"url\":\"classes/Composite.html#dispose\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"Composite\"},{\"kind\":128,\"name\":\"Dataset\",\"url\":\"classes/Dataset.html\",\"classes\":\"tsd-kind-class\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/Dataset.html#constructor\",\"classes\":\"tsd-kind-constructor tsd-parent-kind-class\",\"parent\":\"Dataset\"},{\"kind\":1024,\"name\":\"length\",\"url\":\"classes/Dataset.html#length\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"Dataset\"},{\"kind\":1024,\"name\":\"disposed\",\"url\":\"classes/Dataset.html#disposed\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"Dataset\"},{\"kind\":2048,\"name\":\"update\",\"url\":\"classes/Dataset.html#update\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"Dataset\"},{\"kind\":2048,\"name\":\"dispose\",\"url\":\"classes/Dataset.html#dispose\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"Dataset\"},{\"kind\":2048,\"name\":\"assertFits\",\"url\":\"classes/Dataset.html#assertFits\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-private\",\"parent\":\"Dataset\"},{\"kind\":128,\"name\":\"LinearScale\",\"url\":\"classes/LinearScale.html\",\"classes\":\"tsd-kind-class\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/LinearScale.html#constructor\",\"classes\":\"tsd-kind-constructor tsd-parent-kind-class\",\"parent\":\"LinearScale\"},{\"kind\":1024,\"name\":\"_domain\",\"url\":\"classes/LinearScale.html#_domain\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"LinearScale\"},{\"kind\":1024,\"name\":\"_range\",\"url\":\"classes/LinearScale.html#_range\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"LinearScale\"},{\"kind\":1024,\"name\":\"qDomain\",\"url\":\"classes/LinearScale.html#qDomain\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"LinearScale\"},{\"kind\":1024,\"name\":\"qRange\",\"url\":\"classes/LinearScale.html#qRange\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"LinearScale\"},{\"kind\":262144,\"name\":\"domain\",\"url\":\"classes/LinearScale.html#domain\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class\",\"parent\":\"LinearScale\"},{\"kind\":262144,\"name\":\"range\",\"url\":\"classes/LinearScale.html#range\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class\",\"parent\":\"LinearScale\"},{\"kind\":2048,\"name\":\"toDomain\",\"url\":\"classes/LinearScale.html#toDomain\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"LinearScale\"},{\"kind\":2048,\"name\":\"toRange\",\"url\":\"classes/LinearScale.html#toRange\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"LinearScale\"},{\"kind\":2048,\"name\":\"updateQ\",\"url\":\"classes/LinearScale.html#updateQ\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-private\",\"parent\":\"LinearScale\"},{\"kind\":128,\"name\":\"LogScale\",\"url\":\"classes/LogScale.html\",\"classes\":\"tsd-kind-class\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/LogScale.html#constructor\",\"classes\":\"tsd-kind-constructor tsd-parent-kind-class\",\"parent\":\"LogScale\"},{\"kind\":1024,\"name\":\"_domain\",\"url\":\"classes/LogScale.html#_domain\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"LogScale\"},{\"kind\":1024,\"name\":\"_range\",\"url\":\"classes/LogScale.html#_range\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"LogScale\"},{\"kind\":1024,\"name\":\"conversion\",\"url\":\"classes/LogScale.html#conversion\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"LogScale\"},{\"kind\":1024,\"name\":\"base\",\"url\":\"classes/LogScale.html#base\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"LogScale\"},{\"kind\":262144,\"name\":\"domain\",\"url\":\"classes/LogScale.html#domain\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class\",\"parent\":\"LogScale\"},{\"kind\":262144,\"name\":\"range\",\"url\":\"classes/LogScale.html#range\",\"classes\":\"tsd-kind-accessor tsd-parent-kind-class\",\"parent\":\"LogScale\"},{\"kind\":2048,\"name\":\"toRange\",\"url\":\"classes/LogScale.html#toRange\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"LogScale\"},{\"kind\":2048,\"name\":\"toDomain\",\"url\":\"classes/LogScale.html#toDomain\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"LogScale\"},{\"kind\":128,\"name\":\"CartesianCoordinateSystem\",\"url\":\"classes/CartesianCoordinateSystem.html\",\"classes\":\"tsd-kind-class\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/CartesianCoordinateSystem.html#constructor\",\"classes\":\"tsd-kind-constructor tsd-parent-kind-class\",\"parent\":\"CartesianCoordinateSystem\"},{\"kind\":1024,\"name\":\"xscale\",\"url\":\"classes/CartesianCoordinateSystem.html#xscale\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"CartesianCoordinateSystem\"},{\"kind\":1024,\"name\":\"yscale\",\"url\":\"classes/CartesianCoordinateSystem.html#yscale\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"CartesianCoordinateSystem\"},{\"kind\":2048,\"name\":\"toRange\",\"url\":\"classes/CartesianCoordinateSystem.html#toRange\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"CartesianCoordinateSystem\"},{\"kind\":2048,\"name\":\"toDomain\",\"url\":\"classes/CartesianCoordinateSystem.html#toDomain\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"CartesianCoordinateSystem\"},{\"kind\":128,\"name\":\"PolarCoordinateSystem\",\"url\":\"classes/PolarCoordinateSystem.html\",\"classes\":\"tsd-kind-class\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/PolarCoordinateSystem.html#constructor\",\"classes\":\"tsd-kind-constructor tsd-parent-kind-class\",\"parent\":\"PolarCoordinateSystem\"},{\"kind\":1024,\"name\":\"radialScale\",\"url\":\"classes/PolarCoordinateSystem.html#radialScale\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"PolarCoordinateSystem\"},{\"kind\":1024,\"name\":\"angularScale\",\"url\":\"classes/PolarCoordinateSystem.html#angularScale\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"PolarCoordinateSystem\"},{\"kind\":1024,\"name\":\"xScale\",\"url\":\"classes/PolarCoordinateSystem.html#xScale\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"PolarCoordinateSystem\"},{\"kind\":1024,\"name\":\"yScale\",\"url\":\"classes/PolarCoordinateSystem.html#yScale\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"PolarCoordinateSystem\"},{\"kind\":2048,\"name\":\"toRange\",\"url\":\"classes/PolarCoordinateSystem.html#toRange\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"PolarCoordinateSystem\"},{\"kind\":2048,\"name\":\"toDomain\",\"url\":\"classes/PolarCoordinateSystem.html#toDomain\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"PolarCoordinateSystem\"},{\"kind\":128,\"name\":\"CoordinateSystem\",\"url\":\"classes/CoordinateSystem.html\",\"classes\":\"tsd-kind-class\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/CoordinateSystem.html#constructor\",\"classes\":\"tsd-kind-constructor tsd-parent-kind-class\",\"parent\":\"CoordinateSystem\"},{\"kind\":2048,\"name\":\"toDomain\",\"url\":\"classes/CoordinateSystem.html#toDomain\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"CoordinateSystem\"},{\"kind\":2048,\"name\":\"toRange\",\"url\":\"classes/CoordinateSystem.html#toRange\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"CoordinateSystem\"},{\"kind\":128,\"name\":\"Circles\",\"url\":\"classes/Circles.html\",\"classes\":\"tsd-kind-class\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/Circles.html#constructor\",\"classes\":\"tsd-kind-constructor tsd-parent-kind-class\",\"parent\":\"Circles\"},{\"kind\":1024,\"name\":\"xs\",\"url\":\"classes/Circles.html#xs\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Circles\"},{\"kind\":1024,\"name\":\"ys\",\"url\":\"classes/Circles.html#ys\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Circles\"},{\"kind\":1024,\"name\":\"colors\",\"url\":\"classes/Circles.html#colors\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Circles\"},{\"kind\":1024,\"name\":\"radii\",\"url\":\"classes/Circles.html#radii\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Circles\"},{\"kind\":1024,\"name\":\"borderWidths\",\"url\":\"classes/Circles.html#borderWidths\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Circles\"},{\"kind\":1024,\"name\":\"borderColors\",\"url\":\"classes/Circles.html#borderColors\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Circles\"},{\"kind\":1024,\"name\":\"positionBuffer\",\"url\":\"classes/Circles.html#positionBuffer\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"Circles\"},{\"kind\":2048,\"name\":\"dispose\",\"url\":\"classes/Circles.html#dispose\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"Circles\"},{\"kind\":256,\"name\":\"CirclesOptions\",\"url\":\"interfaces/CirclesOptions.html\",\"classes\":\"tsd-kind-interface\"},{\"kind\":1024,\"name\":\"colors\",\"url\":\"interfaces/CirclesOptions.html#colors\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"CirclesOptions\"},{\"kind\":1024,\"name\":\"radii\",\"url\":\"interfaces/CirclesOptions.html#radii\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"CirclesOptions\"},{\"kind\":1024,\"name\":\"borderColors\",\"url\":\"interfaces/CirclesOptions.html#borderColors\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"CirclesOptions\"},{\"kind\":1024,\"name\":\"borderWidths\",\"url\":\"interfaces/CirclesOptions.html#borderWidths\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"CirclesOptions\"},{\"kind\":128,\"name\":\"HLines\",\"url\":\"classes/HLines.html\",\"classes\":\"tsd-kind-class\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/HLines.html#constructor\",\"classes\":\"tsd-kind-constructor tsd-parent-kind-class\",\"parent\":\"HLines\"},{\"kind\":1024,\"name\":\"lines\",\"url\":\"classes/HLines.html#lines\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"HLines\"},{\"kind\":1024,\"name\":\"widths\",\"url\":\"classes/HLines.html#widths\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"HLines\"},{\"kind\":1024,\"name\":\"colors\",\"url\":\"classes/HLines.html#colors\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"HLines\"},{\"kind\":1024,\"name\":\"segmentGeometry\",\"url\":\"classes/HLines.html#segmentGeometry\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"HLines\"},{\"kind\":2048,\"name\":\"dispose\",\"url\":\"classes/HLines.html#dispose\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"HLines\"},{\"kind\":256,\"name\":\"HLinesOptions\",\"url\":\"interfaces/HLinesOptions.html\",\"classes\":\"tsd-kind-interface\"},{\"kind\":1024,\"name\":\"widths\",\"url\":\"interfaces/HLinesOptions.html#widths\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"HLinesOptions\"},{\"kind\":1024,\"name\":\"colors\",\"url\":\"interfaces/HLinesOptions.html#colors\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"HLinesOptions\"},{\"kind\":128,\"name\":\"InterleavedCircles\",\"url\":\"classes/InterleavedCircles.html\",\"classes\":\"tsd-kind-class\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/InterleavedCircles.html#constructor\",\"classes\":\"tsd-kind-constructor tsd-parent-kind-class\",\"parent\":\"InterleavedCircles\"},{\"kind\":1024,\"name\":\"xys\",\"url\":\"classes/InterleavedCircles.html#xys\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"InterleavedCircles\"},{\"kind\":1024,\"name\":\"colors\",\"url\":\"classes/InterleavedCircles.html#colors\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"InterleavedCircles\"},{\"kind\":1024,\"name\":\"radii\",\"url\":\"classes/InterleavedCircles.html#radii\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"InterleavedCircles\"},{\"kind\":1024,\"name\":\"borderWidths\",\"url\":\"classes/InterleavedCircles.html#borderWidths\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"InterleavedCircles\"},{\"kind\":1024,\"name\":\"borderColors\",\"url\":\"classes/InterleavedCircles.html#borderColors\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"InterleavedCircles\"},{\"kind\":1024,\"name\":\"positionBuffer\",\"url\":\"classes/InterleavedCircles.html#positionBuffer\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"InterleavedCircles\"},{\"kind\":2048,\"name\":\"dispose\",\"url\":\"classes/InterleavedCircles.html#dispose\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"InterleavedCircles\"},{\"kind\":256,\"name\":\"InterleavedCirclesOptions\",\"url\":\"interfaces/InterleavedCirclesOptions.html\",\"classes\":\"tsd-kind-interface\"},{\"kind\":1024,\"name\":\"colors\",\"url\":\"interfaces/InterleavedCirclesOptions.html#colors\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"InterleavedCirclesOptions\"},{\"kind\":1024,\"name\":\"borderColors\",\"url\":\"interfaces/InterleavedCirclesOptions.html#borderColors\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"InterleavedCirclesOptions\"},{\"kind\":1024,\"name\":\"radii\",\"url\":\"interfaces/InterleavedCirclesOptions.html#radii\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"InterleavedCirclesOptions\"},{\"kind\":1024,\"name\":\"borderWidths\",\"url\":\"interfaces/InterleavedCirclesOptions.html#borderWidths\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"InterleavedCirclesOptions\"},{\"kind\":128,\"name\":\"InterleavedShapes\",\"url\":\"classes/InterleavedShapes.html\",\"classes\":\"tsd-kind-class\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/InterleavedShapes.html#constructor\",\"classes\":\"tsd-kind-constructor tsd-parent-kind-class\",\"parent\":\"InterleavedShapes\"},{\"kind\":1024,\"name\":\"shape\",\"url\":\"classes/InterleavedShapes.html#shape\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"InterleavedShapes\"},{\"kind\":1024,\"name\":\"xys\",\"url\":\"classes/InterleavedShapes.html#xys\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"InterleavedShapes\"},{\"kind\":1024,\"name\":\"scales\",\"url\":\"classes/InterleavedShapes.html#scales\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"InterleavedShapes\"},{\"kind\":1024,\"name\":\"rotations\",\"url\":\"classes/InterleavedShapes.html#rotations\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"InterleavedShapes\"},{\"kind\":1024,\"name\":\"colors\",\"url\":\"classes/InterleavedShapes.html#colors\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"InterleavedShapes\"},{\"kind\":2048,\"name\":\"dispose\",\"url\":\"classes/InterleavedShapes.html#dispose\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"InterleavedShapes\"},{\"kind\":256,\"name\":\"InterleavedShapesOptions\",\"url\":\"interfaces/InterleavedShapesOptions.html\",\"classes\":\"tsd-kind-interface\"},{\"kind\":1024,\"name\":\"colors\",\"url\":\"interfaces/InterleavedShapesOptions.html#colors\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"InterleavedShapesOptions\"},{\"kind\":1024,\"name\":\"scales\",\"url\":\"interfaces/InterleavedShapesOptions.html#scales\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"InterleavedShapesOptions\"},{\"kind\":1024,\"name\":\"rotations\",\"url\":\"interfaces/InterleavedShapesOptions.html#rotations\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"InterleavedShapesOptions\"},{\"kind\":128,\"name\":\"LineSegments\",\"url\":\"classes/LineSegments.html\",\"classes\":\"tsd-kind-class\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/LineSegments.html#constructor\",\"classes\":\"tsd-kind-constructor tsd-parent-kind-class\",\"parent\":\"LineSegments\"},{\"kind\":1024,\"name\":\"points\",\"url\":\"classes/LineSegments.html#points\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"LineSegments\"},{\"kind\":1024,\"name\":\"widths\",\"url\":\"classes/LineSegments.html#widths\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"LineSegments\"},{\"kind\":1024,\"name\":\"colors\",\"url\":\"classes/LineSegments.html#colors\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"LineSegments\"},{\"kind\":1024,\"name\":\"segmentGeometry\",\"url\":\"classes/LineSegments.html#segmentGeometry\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"LineSegments\"},{\"kind\":2048,\"name\":\"dispose\",\"url\":\"classes/LineSegments.html#dispose\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"LineSegments\"},{\"kind\":256,\"name\":\"LineSegmentsOptions\",\"url\":\"interfaces/LineSegmentsOptions.html\",\"classes\":\"tsd-kind-interface\"},{\"kind\":1024,\"name\":\"widths\",\"url\":\"interfaces/LineSegmentsOptions.html#widths\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"LineSegmentsOptions\"},{\"kind\":1024,\"name\":\"colors\",\"url\":\"interfaces/LineSegmentsOptions.html#colors\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"LineSegmentsOptions\"},{\"kind\":128,\"name\":\"LineStrip\",\"url\":\"classes/LineStrip.html\",\"classes\":\"tsd-kind-class\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/LineStrip.html#constructor\",\"classes\":\"tsd-kind-constructor tsd-parent-kind-class\",\"parent\":\"LineStrip\"},{\"kind\":1024,\"name\":\"xs\",\"url\":\"classes/LineStrip.html#xs\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"LineStrip\"},{\"kind\":1024,\"name\":\"ys\",\"url\":\"classes/LineStrip.html#ys\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"LineStrip\"},{\"kind\":1024,\"name\":\"widths\",\"url\":\"classes/LineStrip.html#widths\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"LineStrip\"},{\"kind\":1024,\"name\":\"colors\",\"url\":\"classes/LineStrip.html#colors\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"LineStrip\"},{\"kind\":1024,\"name\":\"roundCapJoin\",\"url\":\"classes/LineStrip.html#roundCapJoin\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"LineStrip\"},{\"kind\":2048,\"name\":\"dispose\",\"url\":\"classes/LineStrip.html#dispose\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"LineStrip\"},{\"kind\":256,\"name\":\"LineStripOptions\",\"url\":\"interfaces/LineStripOptions.html\",\"classes\":\"tsd-kind-interface\"},{\"kind\":1024,\"name\":\"widths\",\"url\":\"interfaces/LineStripOptions.html#widths\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"LineStripOptions\"},{\"kind\":1024,\"name\":\"colors\",\"url\":\"interfaces/LineStripOptions.html#colors\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"LineStripOptions\"},{\"kind\":128,\"name\":\"TransparentLineStrip\",\"url\":\"classes/TransparentLineStrip.html\",\"classes\":\"tsd-kind-class\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/TransparentLineStrip.html#constructor\",\"classes\":\"tsd-kind-constructor tsd-parent-kind-class\",\"parent\":\"TransparentLineStrip\"},{\"kind\":1024,\"name\":\"xs\",\"url\":\"classes/TransparentLineStrip.html#xs\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"TransparentLineStrip\"},{\"kind\":1024,\"name\":\"ys\",\"url\":\"classes/TransparentLineStrip.html#ys\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"TransparentLineStrip\"},{\"kind\":1024,\"name\":\"width\",\"url\":\"classes/TransparentLineStrip.html#width\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"TransparentLineStrip\"},{\"kind\":1024,\"name\":\"color\",\"url\":\"classes/TransparentLineStrip.html#color\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"TransparentLineStrip\"},{\"kind\":1024,\"name\":\"segmentPositions\",\"url\":\"classes/TransparentLineStrip.html#segmentPositions\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"TransparentLineStrip\"},{\"kind\":1024,\"name\":\"segmentCells\",\"url\":\"classes/TransparentLineStrip.html#segmentCells\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"TransparentLineStrip\"},{\"kind\":1024,\"name\":\"joinIds\",\"url\":\"classes/TransparentLineStrip.html#joinIds\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"TransparentLineStrip\"},{\"kind\":1024,\"name\":\"joinCells\",\"url\":\"classes/TransparentLineStrip.html#joinCells\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"TransparentLineStrip\"},{\"kind\":1024,\"name\":\"capPositions\",\"url\":\"classes/TransparentLineStrip.html#capPositions\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"TransparentLineStrip\"},{\"kind\":1024,\"name\":\"capCells\",\"url\":\"classes/TransparentLineStrip.html#capCells\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"TransparentLineStrip\"},{\"kind\":2048,\"name\":\"dispose\",\"url\":\"classes/TransparentLineStrip.html#dispose\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"TransparentLineStrip\"},{\"kind\":256,\"name\":\"TransparentLineStripOptions\",\"url\":\"interfaces/TransparentLineStripOptions.html\",\"classes\":\"tsd-kind-interface\"},{\"kind\":1024,\"name\":\"width\",\"url\":\"interfaces/TransparentLineStripOptions.html#width\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"TransparentLineStripOptions\"},{\"kind\":1024,\"name\":\"color\",\"url\":\"interfaces/TransparentLineStripOptions.html#color\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"TransparentLineStripOptions\"},{\"kind\":128,\"name\":\"Rects\",\"url\":\"classes/Rects.html\",\"classes\":\"tsd-kind-class\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/Rects.html#constructor\",\"classes\":\"tsd-kind-constructor tsd-parent-kind-class\",\"parent\":\"Rects\"},{\"kind\":1024,\"name\":\"rects\",\"url\":\"classes/Rects.html#rects\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Rects\"},{\"kind\":1024,\"name\":\"colors\",\"url\":\"classes/Rects.html#colors\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Rects\"},{\"kind\":1024,\"name\":\"positionBuffer\",\"url\":\"classes/Rects.html#positionBuffer\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"Rects\"},{\"kind\":2048,\"name\":\"dispose\",\"url\":\"classes/Rects.html#dispose\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"Rects\"},{\"kind\":256,\"name\":\"RectsOptions\",\"url\":\"interfaces/RectsOptions.html\",\"classes\":\"tsd-kind-interface\"},{\"kind\":1024,\"name\":\"colors\",\"url\":\"interfaces/RectsOptions.html#colors\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"RectsOptions\"},{\"kind\":128,\"name\":\"Shapes\",\"url\":\"classes/Shapes.html\",\"classes\":\"tsd-kind-class\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/Shapes.html#constructor\",\"classes\":\"tsd-kind-constructor tsd-parent-kind-class\",\"parent\":\"Shapes\"},{\"kind\":1024,\"name\":\"shape\",\"url\":\"classes/Shapes.html#shape\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Shapes\"},{\"kind\":1024,\"name\":\"xs\",\"url\":\"classes/Shapes.html#xs\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Shapes\"},{\"kind\":1024,\"name\":\"ys\",\"url\":\"classes/Shapes.html#ys\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Shapes\"},{\"kind\":1024,\"name\":\"scales\",\"url\":\"classes/Shapes.html#scales\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Shapes\"},{\"kind\":1024,\"name\":\"rotations\",\"url\":\"classes/Shapes.html#rotations\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Shapes\"},{\"kind\":1024,\"name\":\"colors\",\"url\":\"classes/Shapes.html#colors\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Shapes\"},{\"kind\":2048,\"name\":\"dispose\",\"url\":\"classes/Shapes.html#dispose\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"Shapes\"},{\"kind\":256,\"name\":\"ShapesOptions\",\"url\":\"interfaces/ShapesOptions.html\",\"classes\":\"tsd-kind-interface\"},{\"kind\":1024,\"name\":\"colors\",\"url\":\"interfaces/ShapesOptions.html#colors\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"ShapesOptions\"},{\"kind\":1024,\"name\":\"scales\",\"url\":\"interfaces/ShapesOptions.html#scales\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"ShapesOptions\"},{\"kind\":1024,\"name\":\"rotations\",\"url\":\"interfaces/ShapesOptions.html#rotations\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"ShapesOptions\"},{\"kind\":128,\"name\":\"Text\",\"url\":\"classes/Text.html\",\"classes\":\"tsd-kind-class\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/Text.html#constructor\",\"classes\":\"tsd-kind-constructor tsd-parent-kind-class\",\"parent\":\"Text\"},{\"kind\":1024,\"name\":\"position\",\"url\":\"classes/Text.html#position\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Text\"},{\"kind\":1024,\"name\":\"size\",\"url\":\"classes/Text.html#size\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Text\"},{\"kind\":1024,\"name\":\"angle\",\"url\":\"classes/Text.html#angle\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Text\"},{\"kind\":1024,\"name\":\"color\",\"url\":\"classes/Text.html#color\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Text\"},{\"kind\":1024,\"name\":\"width\",\"url\":\"classes/Text.html#width\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Text\"},{\"kind\":1024,\"name\":\"height\",\"url\":\"classes/Text.html#height\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Text\"},{\"kind\":1024,\"name\":\"quad\",\"url\":\"classes/Text.html#quad\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"Text\"},{\"kind\":1024,\"name\":\"uv\",\"url\":\"classes/Text.html#uv\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"Text\"},{\"kind\":1024,\"name\":\"instances\",\"url\":\"classes/Text.html#instances\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"Text\"},{\"kind\":1024,\"name\":\"quadGeometry\",\"url\":\"classes/Text.html#quadGeometry\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"Text\"},{\"kind\":2048,\"name\":\"dispose\",\"url\":\"classes/Text.html#dispose\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"Text\"},{\"kind\":256,\"name\":\"TextOptions\",\"url\":\"interfaces/TextOptions.html\",\"classes\":\"tsd-kind-interface\"},{\"kind\":1024,\"name\":\"align\",\"url\":\"interfaces/TextOptions.html#align\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"TextOptions\"},{\"kind\":1024,\"name\":\"anchor\",\"url\":\"interfaces/TextOptions.html#anchor\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"TextOptions\"},{\"kind\":1024,\"name\":\"angle\",\"url\":\"interfaces/TextOptions.html#angle\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"TextOptions\"},{\"kind\":1024,\"name\":\"color\",\"url\":\"interfaces/TextOptions.html#color\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"TextOptions\"},{\"kind\":1024,\"name\":\"size\",\"url\":\"interfaces/TextOptions.html#size\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"TextOptions\"},{\"kind\":128,\"name\":\"Triangles\",\"url\":\"classes/Triangles.html\",\"classes\":\"tsd-kind-class\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/Triangles.html#constructor\",\"classes\":\"tsd-kind-constructor tsd-parent-kind-class\",\"parent\":\"Triangles\"},{\"kind\":1024,\"name\":\"vertices\",\"url\":\"classes/Triangles.html#vertices\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"Triangles\"},{\"kind\":1024,\"name\":\"color\",\"url\":\"classes/Triangles.html#color\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Triangles\"},{\"kind\":2048,\"name\":\"dispose\",\"url\":\"classes/Triangles.html#dispose\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"Triangles\"},{\"kind\":256,\"name\":\"TrianglesOptions\",\"url\":\"interfaces/TrianglesOptions.html\",\"classes\":\"tsd-kind-interface\"},{\"kind\":1024,\"name\":\"color\",\"url\":\"interfaces/TrianglesOptions.html#color\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"TrianglesOptions\"},{\"kind\":128,\"name\":\"VLines\",\"url\":\"classes/VLines.html\",\"classes\":\"tsd-kind-class\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/VLines.html#constructor\",\"classes\":\"tsd-kind-constructor tsd-parent-kind-class\",\"parent\":\"VLines\"},{\"kind\":1024,\"name\":\"lines\",\"url\":\"classes/VLines.html#lines\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"VLines\"},{\"kind\":1024,\"name\":\"widths\",\"url\":\"classes/VLines.html#widths\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"VLines\"},{\"kind\":1024,\"name\":\"colors\",\"url\":\"classes/VLines.html#colors\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"VLines\"},{\"kind\":1024,\"name\":\"segmentGeometry\",\"url\":\"classes/VLines.html#segmentGeometry\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"VLines\"},{\"kind\":2048,\"name\":\"dispose\",\"url\":\"classes/VLines.html#dispose\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"VLines\"},{\"kind\":256,\"name\":\"VLinesOptions\",\"url\":\"interfaces/VLinesOptions.html\",\"classes\":\"tsd-kind-interface\"},{\"kind\":1024,\"name\":\"widths\",\"url\":\"interfaces/VLinesOptions.html#widths\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"VLinesOptions\"},{\"kind\":1024,\"name\":\"colors\",\"url\":\"interfaces/VLinesOptions.html#colors\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"VLinesOptions\"},{\"kind\":128,\"name\":\"Wedges\",\"url\":\"classes/Wedges.html\",\"classes\":\"tsd-kind-class\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/Wedges.html#constructor\",\"classes\":\"tsd-kind-constructor tsd-parent-kind-class\",\"parent\":\"Wedges\"},{\"kind\":1024,\"name\":\"xys\",\"url\":\"classes/Wedges.html#xys\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Wedges\"},{\"kind\":1024,\"name\":\"angles\",\"url\":\"classes/Wedges.html#angles\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Wedges\"},{\"kind\":1024,\"name\":\"colors\",\"url\":\"classes/Wedges.html#colors\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Wedges\"},{\"kind\":1024,\"name\":\"radii\",\"url\":\"classes/Wedges.html#radii\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Wedges\"},{\"kind\":1024,\"name\":\"positionBuffer\",\"url\":\"classes/Wedges.html#positionBuffer\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"Wedges\"},{\"kind\":2048,\"name\":\"dispose\",\"url\":\"classes/Wedges.html#dispose\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"Wedges\"},{\"kind\":256,\"name\":\"WedgesOptions\",\"url\":\"interfaces/WedgesOptions.html\",\"classes\":\"tsd-kind-interface\"},{\"kind\":1024,\"name\":\"colors\",\"url\":\"interfaces/WedgesOptions.html#colors\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"WedgesOptions\"},{\"kind\":1024,\"name\":\"radii\",\"url\":\"interfaces/WedgesOptions.html#radii\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"WedgesOptions\"},{\"kind\":128,\"name\":\"Axis\",\"url\":\"classes/Axis.html\",\"classes\":\"tsd-kind-class\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/Axis.html#constructor\",\"classes\":\"tsd-kind-constructor tsd-parent-kind-class\",\"parent\":\"Axis\"},{\"kind\":1024,\"name\":\"axis\",\"url\":\"classes/Axis.html#axis\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"Axis\"},{\"kind\":1024,\"name\":\"texts\",\"url\":\"classes/Axis.html#texts\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"Axis\"},{\"kind\":1024,\"name\":\"ticks\",\"url\":\"classes/Axis.html#ticks\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"Axis\"},{\"kind\":1024,\"name\":\"minorTicks\",\"url\":\"classes/Axis.html#minorTicks\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"Axis\"},{\"kind\":2048,\"name\":\"dispose\",\"url\":\"classes/Axis.html#dispose\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Axis\"},{\"kind\":256,\"name\":\"AxisOptions\",\"url\":\"interfaces/AxisOptions.html\",\"classes\":\"tsd-kind-interface\"},{\"kind\":1024,\"name\":\"axisColor\",\"url\":\"interfaces/AxisOptions.html#axisColor\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"AxisOptions\"},{\"kind\":1024,\"name\":\"axisWidth\",\"url\":\"interfaces/AxisOptions.html#axisWidth\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"AxisOptions\"},{\"kind\":1024,\"name\":\"labelAnchor\",\"url\":\"interfaces/AxisOptions.html#labelAnchor\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"AxisOptions\"},{\"kind\":1024,\"name\":\"labelAngle\",\"url\":\"interfaces/AxisOptions.html#labelAngle\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"AxisOptions\"},{\"kind\":1024,\"name\":\"labelColor\",\"url\":\"interfaces/AxisOptions.html#labelColor\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"AxisOptions\"},{\"kind\":1024,\"name\":\"labelPad\",\"url\":\"interfaces/AxisOptions.html#labelPad\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"AxisOptions\"},{\"kind\":1024,\"name\":\"labelSide\",\"url\":\"interfaces/AxisOptions.html#labelSide\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"AxisOptions\"},{\"kind\":1024,\"name\":\"labelSize\",\"url\":\"interfaces/AxisOptions.html#labelSize\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"AxisOptions\"},{\"kind\":1024,\"name\":\"tickColor\",\"url\":\"interfaces/AxisOptions.html#tickColor\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"AxisOptions\"},{\"kind\":1024,\"name\":\"tickLength\",\"url\":\"interfaces/AxisOptions.html#tickLength\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"AxisOptions\"},{\"kind\":1024,\"name\":\"tickOffset\",\"url\":\"interfaces/AxisOptions.html#tickOffset\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"AxisOptions\"},{\"kind\":1024,\"name\":\"tickWidth\",\"url\":\"interfaces/AxisOptions.html#tickWidth\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"AxisOptions\"},{\"kind\":1024,\"name\":\"minorTicks\",\"url\":\"interfaces/AxisOptions.html#minorTicks\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"AxisOptions\"},{\"kind\":1024,\"name\":\"minorTickColor\",\"url\":\"interfaces/AxisOptions.html#minorTickColor\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"AxisOptions\"},{\"kind\":1024,\"name\":\"minorTickLength\",\"url\":\"interfaces/AxisOptions.html#minorTickLength\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"AxisOptions\"},{\"kind\":1024,\"name\":\"minorTickOffset\",\"url\":\"interfaces/AxisOptions.html#minorTickOffset\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"AxisOptions\"},{\"kind\":1024,\"name\":\"minorTickWidth\",\"url\":\"interfaces/AxisOptions.html#minorTickWidth\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"AxisOptions\"},{\"kind\":128,\"name\":\"Grid\",\"url\":\"classes/Grid.html\",\"classes\":\"tsd-kind-class\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/Grid.html#constructor\",\"classes\":\"tsd-kind-constructor tsd-parent-kind-class\",\"parent\":\"Grid\"},{\"kind\":1024,\"name\":\"grid\",\"url\":\"classes/Grid.html#grid\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"Grid\"},{\"kind\":2048,\"name\":\"dispose\",\"url\":\"classes/Grid.html#dispose\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Grid\"},{\"kind\":256,\"name\":\"GridOptions\",\"url\":\"interfaces/GridOptions.html\",\"classes\":\"tsd-kind-interface\"},{\"kind\":1024,\"name\":\"width\",\"url\":\"interfaces/GridOptions.html#width\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"GridOptions\"},{\"kind\":1024,\"name\":\"color\",\"url\":\"interfaces/GridOptions.html#color\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"GridOptions\"},{\"kind\":128,\"name\":\"OrthoAxis\",\"url\":\"classes/OrthoAxis.html\",\"classes\":\"tsd-kind-class\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/OrthoAxis.html#constructor\",\"classes\":\"tsd-kind-constructor tsd-parent-kind-class\",\"parent\":\"OrthoAxis\"},{\"kind\":1024,\"name\":\"info\",\"url\":\"classes/OrthoAxis.html#info\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"OrthoAxis\"},{\"kind\":1024,\"name\":\"axis\",\"url\":\"classes/OrthoAxis.html#axis\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"OrthoAxis\"},{\"kind\":2048,\"name\":\"dispose\",\"url\":\"classes/OrthoAxis.html#dispose\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"OrthoAxis\"},{\"kind\":256,\"name\":\"OrthoAxisOptions\",\"url\":\"interfaces/OrthoAxisOptions.html\",\"classes\":\"tsd-kind-interface\"},{\"kind\":1024,\"name\":\"axisHigh\",\"url\":\"interfaces/OrthoAxisOptions.html#axisHigh\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"OrthoAxisOptions\"},{\"kind\":1024,\"name\":\"axisIntercept\",\"url\":\"interfaces/OrthoAxisOptions.html#axisIntercept\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"OrthoAxisOptions\"},{\"kind\":1024,\"name\":\"axisLow\",\"url\":\"interfaces/OrthoAxisOptions.html#axisLow\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"OrthoAxisOptions\"},{\"kind\":1024,\"name\":\"labelFormatter\",\"url\":\"interfaces/OrthoAxisOptions.html#labelFormatter\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"OrthoAxisOptions\"},{\"kind\":65536,\"name\":\"__type\",\"url\":\"interfaces/OrthoAxisOptions.html#__type\",\"classes\":\"tsd-kind-type-literal tsd-parent-kind-interface\",\"parent\":\"OrthoAxisOptions\"},{\"kind\":1024,\"name\":\"minorTickCount\",\"url\":\"interfaces/OrthoAxisOptions.html#minorTickCount\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"OrthoAxisOptions\"},{\"kind\":1024,\"name\":\"tickOrigin\",\"url\":\"interfaces/OrthoAxisOptions.html#tickOrigin\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"OrthoAxisOptions\"},{\"kind\":1024,\"name\":\"tickStep\",\"url\":\"interfaces/OrthoAxisOptions.html#tickStep\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"OrthoAxisOptions\"},{\"kind\":1024,\"name\":\"axisColor\",\"url\":\"interfaces/OrthoAxisOptions.html#axisColor\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface tsd-is-inherited\",\"parent\":\"OrthoAxisOptions\"},{\"kind\":1024,\"name\":\"axisWidth\",\"url\":\"interfaces/OrthoAxisOptions.html#axisWidth\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface tsd-is-inherited\",\"parent\":\"OrthoAxisOptions\"},{\"kind\":1024,\"name\":\"labelAnchor\",\"url\":\"interfaces/OrthoAxisOptions.html#labelAnchor\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface tsd-is-inherited\",\"parent\":\"OrthoAxisOptions\"},{\"kind\":1024,\"name\":\"labelAngle\",\"url\":\"interfaces/OrthoAxisOptions.html#labelAngle\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface tsd-is-inherited\",\"parent\":\"OrthoAxisOptions\"},{\"kind\":1024,\"name\":\"labelColor\",\"url\":\"interfaces/OrthoAxisOptions.html#labelColor\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface tsd-is-inherited\",\"parent\":\"OrthoAxisOptions\"},{\"kind\":1024,\"name\":\"labelPad\",\"url\":\"interfaces/OrthoAxisOptions.html#labelPad\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface tsd-is-inherited\",\"parent\":\"OrthoAxisOptions\"},{\"kind\":1024,\"name\":\"labelSide\",\"url\":\"interfaces/OrthoAxisOptions.html#labelSide\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface tsd-is-inherited\",\"parent\":\"OrthoAxisOptions\"},{\"kind\":1024,\"name\":\"labelSize\",\"url\":\"interfaces/OrthoAxisOptions.html#labelSize\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface tsd-is-inherited\",\"parent\":\"OrthoAxisOptions\"},{\"kind\":1024,\"name\":\"tickColor\",\"url\":\"interfaces/OrthoAxisOptions.html#tickColor\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface tsd-is-inherited\",\"parent\":\"OrthoAxisOptions\"},{\"kind\":1024,\"name\":\"tickLength\",\"url\":\"interfaces/OrthoAxisOptions.html#tickLength\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface tsd-is-inherited\",\"parent\":\"OrthoAxisOptions\"},{\"kind\":1024,\"name\":\"tickOffset\",\"url\":\"interfaces/OrthoAxisOptions.html#tickOffset\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface tsd-is-inherited\",\"parent\":\"OrthoAxisOptions\"},{\"kind\":1024,\"name\":\"tickWidth\",\"url\":\"interfaces/OrthoAxisOptions.html#tickWidth\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface tsd-is-inherited\",\"parent\":\"OrthoAxisOptions\"},{\"kind\":1024,\"name\":\"minorTicks\",\"url\":\"interfaces/OrthoAxisOptions.html#minorTicks\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface tsd-is-inherited\",\"parent\":\"OrthoAxisOptions\"},{\"kind\":1024,\"name\":\"minorTickColor\",\"url\":\"interfaces/OrthoAxisOptions.html#minorTickColor\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface tsd-is-inherited\",\"parent\":\"OrthoAxisOptions\"},{\"kind\":1024,\"name\":\"minorTickLength\",\"url\":\"interfaces/OrthoAxisOptions.html#minorTickLength\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface tsd-is-inherited\",\"parent\":\"OrthoAxisOptions\"},{\"kind\":1024,\"name\":\"minorTickOffset\",\"url\":\"interfaces/OrthoAxisOptions.html#minorTickOffset\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface tsd-is-inherited\",\"parent\":\"OrthoAxisOptions\"},{\"kind\":1024,\"name\":\"minorTickWidth\",\"url\":\"interfaces/OrthoAxisOptions.html#minorTickWidth\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface tsd-is-inherited\",\"parent\":\"OrthoAxisOptions\"},{\"kind\":256,\"name\":\"OrthoAxisInfo\",\"url\":\"interfaces/OrthoAxisInfo.html\",\"classes\":\"tsd-kind-interface\"},{\"kind\":1024,\"name\":\"ticks\",\"url\":\"interfaces/OrthoAxisInfo.html#ticks\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"OrthoAxisInfo\"},{\"kind\":1024,\"name\":\"minorTicks\",\"url\":\"interfaces/OrthoAxisInfo.html#minorTicks\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"OrthoAxisInfo\"},{\"kind\":128,\"name\":\"Scissor\",\"url\":\"classes/Scissor.html\",\"classes\":\"tsd-kind-class\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/Scissor.html#constructor\",\"classes\":\"tsd-kind-constructor tsd-parent-kind-class\",\"parent\":\"Scissor\"},{\"kind\":1024,\"name\":\"_children\",\"url\":\"classes/Scissor.html#_children\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"Scissor\"},{\"kind\":1024,\"name\":\"scope\",\"url\":\"classes/Scissor.html#scope\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Scissor\"},{\"kind\":1024,\"name\":\"x\",\"url\":\"classes/Scissor.html#x\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Scissor\"},{\"kind\":1024,\"name\":\"y\",\"url\":\"classes/Scissor.html#y\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Scissor\"},{\"kind\":1024,\"name\":\"width\",\"url\":\"classes/Scissor.html#width\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Scissor\"},{\"kind\":1024,\"name\":\"height\",\"url\":\"classes/Scissor.html#height\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Scissor\"},{\"kind\":1024,\"name\":\"screenSpace\",\"url\":\"classes/Scissor.html#screenSpace\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Scissor\"},{\"kind\":2048,\"name\":\"dispose\",\"url\":\"classes/Scissor.html#dispose\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"Scissor\"},{\"kind\":64,\"name\":\"createDefaultFont\",\"url\":\"functions/createDefaultFont.html\",\"classes\":\"tsd-kind-function\"},{\"kind\":128,\"name\":\"Font\",\"url\":\"classes/Font.html\",\"classes\":\"tsd-kind-class\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/Font.html#constructor\",\"classes\":\"tsd-kind-constructor tsd-parent-kind-class\",\"parent\":\"Font\"},{\"kind\":1024,\"name\":\"kernTable\",\"url\":\"classes/Font.html#kernTable\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"Font\"},{\"kind\":1024,\"name\":\"maxid\",\"url\":\"classes/Font.html#maxid\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-private\",\"parent\":\"Font\"},{\"kind\":2048,\"name\":\"dispose\",\"url\":\"classes/Font.html#dispose\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"Font\"},{\"kind\":8388608,\"name\":\"default\",\"url\":\"index.html#default\",\"classes\":\"tsd-kind-reference\"}],\"index\":{\"version\":\"2.3.9\",\"fields\":[\"name\",\"comment\"],\"fieldVectors\":[[\"name/0\",[0,53.534]],[\"comment/0\",[]],[\"name/1\",[1,24.447]],[\"comment/1\",[]],[\"name/2\",[2,48.426]],[\"comment/2\",[]],[\"name/3\",[3,53.534]],[\"comment/3\",[]],[\"name/4\",[4,48.426]],[\"comment/4\",[]],[\"name/5\",[5,53.534]],[\"comment/5\",[]],[\"name/6\",[6,53.534]],[\"comment/6\",[]],[\"name/7\",[7,53.534]],[\"comment/7\",[]],[\"name/8\",[8,53.534]],[\"comment/8\",[]],[\"name/9\",[9,53.534]],[\"comment/9\",[]],[\"name/10\",[10,53.534]],[\"comment/10\",[]],[\"name/11\",[11,53.534]],[\"comment/11\",[]],[\"name/12\",[2,48.426]],[\"comment/12\",[]],[\"name/13\",[12,53.534]],[\"comment/13\",[]],[\"name/14\",[13,53.534]],[\"comment/14\",[]],[\"name/15\",[14,48.426]],[\"comment/15\",[]],[\"name/16\",[15,48.426]],[\"comment/16\",[]],[\"name/17\",[16,38.871]],[\"comment/17\",[]],[\"name/18\",[17,45.061]],[\"comment/18\",[]],[\"name/19\",[18,53.534]],[\"comment/19\",[]],[\"name/20\",[19,53.534]],[\"comment/20\",[]],[\"name/21\",[20,53.534]],[\"comment/21\",[]],[\"name/22\",[21,53.534]],[\"comment/22\",[]],[\"name/23\",[22,53.534]],[\"comment/23\",[]],[\"name/24\",[23,53.534]],[\"comment/24\",[]],[\"name/25\",[24,53.534]],[\"comment/25\",[]],[\"name/26\",[1,24.447]],[\"comment/26\",[]],[\"name/27\",[25,26.908]],[\"comment/27\",[]],[\"name/28\",[26,53.534]],[\"comment/28\",[]],[\"name/29\",[1,24.447]],[\"comment/29\",[]],[\"name/30\",[25,26.908]],[\"comment/30\",[]],[\"name/31\",[27,53.534]],[\"comment/31\",[]],[\"name/32\",[1,24.447]],[\"comment/32\",[]],[\"name/33\",[28,53.534]],[\"comment/33\",[]],[\"name/34\",[29,53.534]],[\"comment/34\",[]],[\"name/35\",[30,53.534]],[\"comment/35\",[]],[\"name/36\",[25,26.908]],[\"comment/36\",[]],[\"name/37\",[31,53.534]],[\"comment/37\",[]],[\"name/38\",[32,53.534]],[\"comment/38\",[]],[\"name/39\",[1,24.447]],[\"comment/39\",[]],[\"name/40\",[33,48.426]],[\"comment/40\",[]],[\"name/41\",[34,48.426]],[\"comment/41\",[]],[\"name/42\",[35,53.534]],[\"comment/42\",[]],[\"name/43\",[36,53.534]],[\"comment/43\",[]],[\"name/44\",[37,48.426]],[\"comment/44\",[]],[\"name/45\",[38,48.426]],[\"comment/45\",[]],[\"name/46\",[39,40.542]],[\"comment/46\",[]],[\"name/47\",[40,40.542]],[\"comment/47\",[]],[\"name/48\",[41,53.534]],[\"comment/48\",[]],[\"name/49\",[42,53.534]],[\"comment/49\",[]],[\"name/50\",[1,24.447]],[\"comment/50\",[]],[\"name/51\",[33,48.426]],[\"comment/51\",[]],[\"name/52\",[34,48.426]],[\"comment/52\",[]],[\"name/53\",[43,53.534]],[\"comment/53\",[]],[\"name/54\",[44,53.534]],[\"comment/54\",[]],[\"name/55\",[37,48.426]],[\"comment/55\",[]],[\"name/56\",[38,48.426]],[\"comment/56\",[]],[\"name/57\",[40,40.542]],[\"comment/57\",[]],[\"name/58\",[39,40.542]],[\"comment/58\",[]],[\"name/59\",[45,53.534]],[\"comment/59\",[]],[\"name/60\",[1,24.447]],[\"comment/60\",[]],[\"name/61\",[46,48.426]],[\"comment/61\",[]],[\"name/62\",[47,48.426]],[\"comment/62\",[]],[\"name/63\",[40,40.542]],[\"comment/63\",[]],[\"name/64\",[39,40.542]],[\"comment/64\",[]],[\"name/65\",[48,53.534]],[\"comment/65\",[]],[\"name/66\",[1,24.447]],[\"comment/66\",[]],[\"name/67\",[49,53.534]],[\"comment/67\",[]],[\"name/68\",[50,53.534]],[\"comment/68\",[]],[\"name/69\",[46,48.426]],[\"comment/69\",[]],[\"name/70\",[47,48.426]],[\"comment/70\",[]],[\"name/71\",[40,40.542]],[\"comment/71\",[]],[\"name/72\",[39,40.542]],[\"comment/72\",[]],[\"name/73\",[51,53.534]],[\"comment/73\",[]],[\"name/74\",[1,24.447]],[\"comment/74\",[]],[\"name/75\",[39,40.542]],[\"comment/75\",[]],[\"name/76\",[40,40.542]],[\"comment/76\",[]],[\"name/77\",[52,53.534]],[\"comment/77\",[]],[\"name/78\",[1,24.447]],[\"comment/78\",[]],[\"name/79\",[53,42.548]],[\"comment/79\",[]],[\"name/80\",[54,42.548]],[\"comment/80\",[]],[\"name/81\",[55,27.385]],[\"comment/81\",[]],[\"name/82\",[56,38.871]],[\"comment/82\",[]],[\"name/83\",[57,42.548]],[\"comment/83\",[]],[\"name/84\",[58,42.548]],[\"comment/84\",[]],[\"name/85\",[59,42.548]],[\"comment/85\",[]],[\"name/86\",[25,26.908]],[\"comment/86\",[]],[\"name/87\",[60,53.534]],[\"comment/87\",[]],[\"name/88\",[55,27.385]],[\"comment/88\",[]],[\"name/89\",[56,38.871]],[\"comment/89\",[]],[\"name/90\",[58,42.548]],[\"comment/90\",[]],[\"name/91\",[57,42.548]],[\"comment/91\",[]],[\"name/92\",[61,53.534]],[\"comment/92\",[]],[\"name/93\",[1,24.447]],[\"comment/93\",[]],[\"name/94\",[62,48.426]],[\"comment/94\",[]],[\"name/95\",[63,36.188]],[\"comment/95\",[]],[\"name/96\",[55,27.385]],[\"comment/96\",[]],[\"name/97\",[64,45.061]],[\"comment/97\",[]],[\"name/98\",[25,26.908]],[\"comment/98\",[]],[\"name/99\",[65,53.534]],[\"comment/99\",[]],[\"name/100\",[63,36.188]],[\"comment/100\",[]],[\"name/101\",[55,27.385]],[\"comment/101\",[]],[\"name/102\",[66,53.534]],[\"comment/102\",[]],[\"name/103\",[1,24.447]],[\"comment/103\",[]],[\"name/104\",[67,45.061]],[\"comment/104\",[]],[\"name/105\",[55,27.385]],[\"comment/105\",[]],[\"name/106\",[56,38.871]],[\"comment/106\",[]],[\"name/107\",[57,42.548]],[\"comment/107\",[]],[\"name/108\",[58,42.548]],[\"comment/108\",[]],[\"name/109\",[59,42.548]],[\"comment/109\",[]],[\"name/110\",[25,26.908]],[\"comment/110\",[]],[\"name/111\",[68,53.534]],[\"comment/111\",[]],[\"name/112\",[55,27.385]],[\"comment/112\",[]],[\"name/113\",[58,42.548]],[\"comment/113\",[]],[\"name/114\",[56,38.871]],[\"comment/114\",[]],[\"name/115\",[57,42.548]],[\"comment/115\",[]],[\"name/116\",[69,53.534]],[\"comment/116\",[]],[\"name/117\",[1,24.447]],[\"comment/117\",[]],[\"name/118\",[70,48.426]],[\"comment/118\",[]],[\"name/119\",[67,45.061]],[\"comment/119\",[]],[\"name/120\",[71,42.548]],[\"comment/120\",[]],[\"name/121\",[72,42.548]],[\"comment/121\",[]],[\"name/122\",[55,27.385]],[\"comment/122\",[]],[\"name/123\",[25,26.908]],[\"comment/123\",[]],[\"name/124\",[73,53.534]],[\"comment/124\",[]],[\"name/125\",[55,27.385]],[\"comment/125\",[]],[\"name/126\",[71,42.548]],[\"comment/126\",[]],[\"name/127\",[72,42.548]],[\"comment/127\",[]],[\"name/128\",[74,53.534]],[\"comment/128\",[]],[\"name/129\",[1,24.447]],[\"comment/129\",[]],[\"name/130\",[75,53.534]],[\"comment/130\",[]],[\"name/131\",[63,36.188]],[\"comment/131\",[]],[\"name/132\",[55,27.385]],[\"comment/132\",[]],[\"name/133\",[64,45.061]],[\"comment/133\",[]],[\"name/134\",[25,26.908]],[\"comment/134\",[]],[\"name/135\",[76,53.534]],[\"comment/135\",[]],[\"name/136\",[63,36.188]],[\"comment/136\",[]],[\"name/137\",[55,27.385]],[\"comment/137\",[]],[\"name/138\",[77,53.534]],[\"comment/138\",[]],[\"name/139\",[1,24.447]],[\"comment/139\",[]],[\"name/140\",[53,42.548]],[\"comment/140\",[]],[\"name/141\",[54,42.548]],[\"comment/141\",[]],[\"name/142\",[63,36.188]],[\"comment/142\",[]],[\"name/143\",[55,27.385]],[\"comment/143\",[]],[\"name/144\",[78,53.534]],[\"comment/144\",[]],[\"name/145\",[25,26.908]],[\"comment/145\",[]],[\"name/146\",[79,53.534]],[\"comment/146\",[]],[\"name/147\",[63,36.188]],[\"comment/147\",[]],[\"name/148\",[55,27.385]],[\"comment/148\",[]],[\"name/149\",[80,53.534]],[\"comment/149\",[]],[\"name/150\",[1,24.447]],[\"comment/150\",[]],[\"name/151\",[53,42.548]],[\"comment/151\",[]],[\"name/152\",[54,42.548]],[\"comment/152\",[]],[\"name/153\",[16,38.871]],[\"comment/153\",[]],[\"name/154\",[81,37.44]],[\"comment/154\",[]],[\"name/155\",[82,53.534]],[\"comment/155\",[]],[\"name/156\",[83,53.534]],[\"comment/156\",[]],[\"name/157\",[84,53.534]],[\"comment/157\",[]],[\"name/158\",[85,53.534]],[\"comment/158\",[]],[\"name/159\",[86,53.534]],[\"comment/159\",[]],[\"name/160\",[87,53.534]],[\"comment/160\",[]],[\"name/161\",[25,26.908]],[\"comment/161\",[]],[\"name/162\",[88,53.534]],[\"comment/162\",[]],[\"name/163\",[16,38.871]],[\"comment/163\",[]],[\"name/164\",[81,37.44]],[\"comment/164\",[]],[\"name/165\",[89,48.426]],[\"comment/165\",[]],[\"name/166\",[1,24.447]],[\"comment/166\",[]],[\"name/167\",[89,48.426]],[\"comment/167\",[]],[\"name/168\",[55,27.385]],[\"comment/168\",[]],[\"name/169\",[59,42.548]],[\"comment/169\",[]],[\"name/170\",[25,26.908]],[\"comment/170\",[]],[\"name/171\",[90,53.534]],[\"comment/171\",[]],[\"name/172\",[55,27.385]],[\"comment/172\",[]],[\"name/173\",[91,53.534]],[\"comment/173\",[]],[\"name/174\",[1,24.447]],[\"comment/174\",[]],[\"name/175\",[70,48.426]],[\"comment/175\",[]],[\"name/176\",[53,42.548]],[\"comment/176\",[]],[\"name/177\",[54,42.548]],[\"comment/177\",[]],[\"name/178\",[71,42.548]],[\"comment/178\",[]],[\"name/179\",[72,42.548]],[\"comment/179\",[]],[\"name/180\",[55,27.385]],[\"comment/180\",[]],[\"name/181\",[25,26.908]],[\"comment/181\",[]],[\"name/182\",[92,53.534]],[\"comment/182\",[]],[\"name/183\",[55,27.385]],[\"comment/183\",[]],[\"name/184\",[71,42.548]],[\"comment/184\",[]],[\"name/185\",[72,42.548]],[\"comment/185\",[]],[\"name/186\",[93,53.534]],[\"comment/186\",[]],[\"name/187\",[1,24.447]],[\"comment/187\",[]],[\"name/188\",[94,53.534]],[\"comment/188\",[]],[\"name/189\",[95,48.426]],[\"comment/189\",[]],[\"name/190\",[96,48.426]],[\"comment/190\",[]],[\"name/191\",[81,37.44]],[\"comment/191\",[]],[\"name/192\",[16,38.871]],[\"comment/192\",[]],[\"name/193\",[17,45.061]],[\"comment/193\",[]],[\"name/194\",[97,53.534]],[\"comment/194\",[]],[\"name/195\",[98,53.534]],[\"comment/195\",[]],[\"name/196\",[99,53.534]],[\"comment/196\",[]],[\"name/197\",[100,53.534]],[\"comment/197\",[]],[\"name/198\",[25,26.908]],[\"comment/198\",[]],[\"name/199\",[101,53.534]],[\"comment/199\",[]],[\"name/200\",[102,53.534]],[\"comment/200\",[]],[\"name/201\",[103,53.534]],[\"comment/201\",[]],[\"name/202\",[96,48.426]],[\"comment/202\",[]],[\"name/203\",[81,37.44]],[\"comment/203\",[]],[\"name/204\",[95,48.426]],[\"comment/204\",[]],[\"name/205\",[104,53.534]],[\"comment/205\",[]],[\"name/206\",[1,24.447]],[\"comment/206\",[]],[\"name/207\",[105,53.534]],[\"comment/207\",[]],[\"name/208\",[81,37.44]],[\"comment/208\",[]],[\"name/209\",[25,26.908]],[\"comment/209\",[]],[\"name/210\",[106,53.534]],[\"comment/210\",[]],[\"name/211\",[81,37.44]],[\"comment/211\",[]],[\"name/212\",[107,53.534]],[\"comment/212\",[]],[\"name/213\",[1,24.447]],[\"comment/213\",[]],[\"name/214\",[62,48.426]],[\"comment/214\",[]],[\"name/215\",[63,36.188]],[\"comment/215\",[]],[\"name/216\",[55,27.385]],[\"comment/216\",[]],[\"name/217\",[64,45.061]],[\"comment/217\",[]],[\"name/218\",[25,26.908]],[\"comment/218\",[]],[\"name/219\",[108,53.534]],[\"comment/219\",[]],[\"name/220\",[63,36.188]],[\"comment/220\",[]],[\"name/221\",[55,27.385]],[\"comment/221\",[]],[\"name/222\",[109,53.534]],[\"comment/222\",[]],[\"name/223\",[1,24.447]],[\"comment/223\",[]],[\"name/224\",[67,45.061]],[\"comment/224\",[]],[\"name/225\",[110,53.534]],[\"comment/225\",[]],[\"name/226\",[55,27.385]],[\"comment/226\",[]],[\"name/227\",[56,38.871]],[\"comment/227\",[]],[\"name/228\",[59,42.548]],[\"comment/228\",[]],[\"name/229\",[25,26.908]],[\"comment/229\",[]],[\"name/230\",[111,53.534]],[\"comment/230\",[]],[\"name/231\",[55,27.385]],[\"comment/231\",[]],[\"name/232\",[56,38.871]],[\"comment/232\",[]],[\"name/233\",[112,45.061]],[\"comment/233\",[]],[\"name/234\",[1,24.447]],[\"comment/234\",[]],[\"name/235\",[112,45.061]],[\"comment/235\",[]],[\"name/236\",[113,53.534]],[\"comment/236\",[]],[\"name/237\",[114,48.426]],[\"comment/237\",[]],[\"name/238\",[115,42.548]],[\"comment/238\",[]],[\"name/239\",[25,26.908]],[\"comment/239\",[]],[\"name/240\",[116,53.534]],[\"comment/240\",[]],[\"name/241\",[117,48.426]],[\"comment/241\",[]],[\"name/242\",[118,48.426]],[\"comment/242\",[]],[\"name/243\",[119,48.426]],[\"comment/243\",[]],[\"name/244\",[120,48.426]],[\"comment/244\",[]],[\"name/245\",[121,48.426]],[\"comment/245\",[]],[\"name/246\",[122,48.426]],[\"comment/246\",[]],[\"name/247\",[123,48.426]],[\"comment/247\",[]],[\"name/248\",[124,48.426]],[\"comment/248\",[]],[\"name/249\",[125,48.426]],[\"comment/249\",[]],[\"name/250\",[126,48.426]],[\"comment/250\",[]],[\"name/251\",[127,48.426]],[\"comment/251\",[]],[\"name/252\",[128,48.426]],[\"comment/252\",[]],[\"name/253\",[115,42.548]],[\"comment/253\",[]],[\"name/254\",[129,48.426]],[\"comment/254\",[]],[\"name/255\",[130,48.426]],[\"comment/255\",[]],[\"name/256\",[131,48.426]],[\"comment/256\",[]],[\"name/257\",[132,48.426]],[\"comment/257\",[]],[\"name/258\",[133,48.426]],[\"comment/258\",[]],[\"name/259\",[1,24.447]],[\"comment/259\",[]],[\"name/260\",[133,48.426]],[\"comment/260\",[]],[\"name/261\",[25,26.908]],[\"comment/261\",[]],[\"name/262\",[134,53.534]],[\"comment/262\",[]],[\"name/263\",[16,38.871]],[\"comment/263\",[]],[\"name/264\",[81,37.44]],[\"comment/264\",[]],[\"name/265\",[135,53.534]],[\"comment/265\",[]],[\"name/266\",[1,24.447]],[\"comment/266\",[]],[\"name/267\",[136,53.534]],[\"comment/267\",[]],[\"name/268\",[112,45.061]],[\"comment/268\",[]],[\"name/269\",[25,26.908]],[\"comment/269\",[]],[\"name/270\",[137,53.534]],[\"comment/270\",[]],[\"name/271\",[138,53.534]],[\"comment/271\",[]],[\"name/272\",[139,53.534]],[\"comment/272\",[]],[\"name/273\",[140,53.534]],[\"comment/273\",[]],[\"name/274\",[141,53.534]],[\"comment/274\",[]],[\"name/275\",[142,53.534]],[\"comment/275\",[]],[\"name/276\",[143,53.534]],[\"comment/276\",[]],[\"name/277\",[144,53.534]],[\"comment/277\",[]],[\"name/278\",[145,53.534]],[\"comment/278\",[]],[\"name/279\",[117,48.426]],[\"comment/279\",[]],[\"name/280\",[118,48.426]],[\"comment/280\",[]],[\"name/281\",[119,48.426]],[\"comment/281\",[]],[\"name/282\",[120,48.426]],[\"comment/282\",[]],[\"name/283\",[121,48.426]],[\"comment/283\",[]],[\"name/284\",[122,48.426]],[\"comment/284\",[]],[\"name/285\",[123,48.426]],[\"comment/285\",[]],[\"name/286\",[124,48.426]],[\"comment/286\",[]],[\"name/287\",[125,48.426]],[\"comment/287\",[]],[\"name/288\",[126,48.426]],[\"comment/288\",[]],[\"name/289\",[127,48.426]],[\"comment/289\",[]],[\"name/290\",[128,48.426]],[\"comment/290\",[]],[\"name/291\",[115,42.548]],[\"comment/291\",[]],[\"name/292\",[129,48.426]],[\"comment/292\",[]],[\"name/293\",[130,48.426]],[\"comment/293\",[]],[\"name/294\",[131,48.426]],[\"comment/294\",[]],[\"name/295\",[132,48.426]],[\"comment/295\",[]],[\"name/296\",[146,53.534]],[\"comment/296\",[]],[\"name/297\",[114,48.426]],[\"comment/297\",[]],[\"name/298\",[115,42.548]],[\"comment/298\",[]],[\"name/299\",[147,53.534]],[\"comment/299\",[]],[\"name/300\",[1,24.447]],[\"comment/300\",[]],[\"name/301\",[148,53.534]],[\"comment/301\",[]],[\"name/302\",[4,48.426]],[\"comment/302\",[]],[\"name/303\",[14,48.426]],[\"comment/303\",[]],[\"name/304\",[15,48.426]],[\"comment/304\",[]],[\"name/305\",[16,38.871]],[\"comment/305\",[]],[\"name/306\",[17,45.061]],[\"comment/306\",[]],[\"name/307\",[149,53.534]],[\"comment/307\",[]],[\"name/308\",[25,26.908]],[\"comment/308\",[]],[\"name/309\",[150,53.534]],[\"comment/309\",[]],[\"name/310\",[151,53.534]],[\"comment/310\",[]],[\"name/311\",[1,24.447]],[\"comment/311\",[]],[\"name/312\",[152,53.534]],[\"comment/312\",[]],[\"name/313\",[153,53.534]],[\"comment/313\",[]],[\"name/314\",[25,26.908]],[\"comment/314\",[]],[\"name/315\",[154,53.534]],[\"comment/315\",[]]],\"invertedIndex\":[[\"__type\",{\"_index\":142,\"name\":{\"275\":{}},\"comment\":{}}],[\"_children\",{\"_index\":148,\"name\":{\"301\":{}},\"comment\":{}}],[\"_domain\",{\"_index\":33,\"name\":{\"40\":{},\"51\":{}},\"comment\":{}}],[\"_range\",{\"_index\":34,\"name\":{\"41\":{},\"52\":{}},\"comment\":{}}],[\"align\",{\"_index\":102,\"name\":{\"200\":{}},\"comment\":{}}],[\"alpha\",{\"_index\":12,\"name\":{\"13\":{}},\"comment\":{}}],[\"anchor\",{\"_index\":103,\"name\":{\"201\":{}},\"comment\":{}}],[\"angle\",{\"_index\":96,\"name\":{\"190\":{},\"202\":{}},\"comment\":{}}],[\"angles\",{\"_index\":110,\"name\":{\"225\":{}},\"comment\":{}}],[\"angularscale\",{\"_index\":50,\"name\":{\"68\":{}},\"comment\":{}}],[\"assertfits\",{\"_index\":31,\"name\":{\"37\":{}},\"comment\":{}}],[\"axis\",{\"_index\":112,\"name\":{\"233\":{},\"235\":{},\"268\":{}},\"comment\":{}}],[\"axiscolor\",{\"_index\":117,\"name\":{\"241\":{},\"279\":{}},\"comment\":{}}],[\"axishigh\",{\"_index\":138,\"name\":{\"271\":{}},\"comment\":{}}],[\"axisintercept\",{\"_index\":139,\"name\":{\"272\":{}},\"comment\":{}}],[\"axislow\",{\"_index\":140,\"name\":{\"273\":{}},\"comment\":{}}],[\"axisoptions\",{\"_index\":116,\"name\":{\"240\":{}},\"comment\":{}}],[\"axiswidth\",{\"_index\":118,\"name\":{\"242\":{},\"280\":{}},\"comment\":{}}],[\"base\",{\"_index\":44,\"name\":{\"54\":{}},\"comment\":{}}],[\"bordercolors\",{\"_index\":58,\"name\":{\"84\":{},\"90\":{},\"108\":{},\"113\":{}},\"comment\":{}}],[\"borderwidths\",{\"_index\":57,\"name\":{\"83\":{},\"91\":{},\"107\":{},\"115\":{}},\"comment\":{}}],[\"candygraph\",{\"_index\":0,\"name\":{\"0\":{}},\"comment\":{}}],[\"candygraphoptions\",{\"_index\":11,\"name\":{\"11\":{}},\"comment\":{}}],[\"canvas\",{\"_index\":2,\"name\":{\"2\":{},\"12\":{}},\"comment\":{}}],[\"capcells\",{\"_index\":87,\"name\":{\"160\":{}},\"comment\":{}}],[\"cappositions\",{\"_index\":86,\"name\":{\"159\":{}},\"comment\":{}}],[\"cartesiancoordinatesystem\",{\"_index\":45,\"name\":{\"59\":{}},\"comment\":{}}],[\"circles\",{\"_index\":52,\"name\":{\"77\":{}},\"comment\":{}}],[\"circlesoptions\",{\"_index\":60,\"name\":{\"87\":{}},\"comment\":{}}],[\"clear\",{\"_index\":5,\"name\":{\"5\":{}},\"comment\":{}}],[\"color\",{\"_index\":81,\"name\":{\"154\":{},\"164\":{},\"191\":{},\"203\":{},\"208\":{},\"211\":{},\"264\":{}},\"comment\":{}}],[\"colors\",{\"_index\":55,\"name\":{\"81\":{},\"88\":{},\"96\":{},\"101\":{},\"105\":{},\"112\":{},\"122\":{},\"125\":{},\"132\":{},\"137\":{},\"143\":{},\"148\":{},\"168\":{},\"172\":{},\"180\":{},\"183\":{},\"216\":{},\"221\":{},\"226\":{},\"231\":{}},\"comment\":{}}],[\"commandsbycoords\",{\"_index\":3,\"name\":{\"3\":{}},\"comment\":{}}],[\"composite\",{\"_index\":26,\"name\":{\"28\":{}},\"comment\":{}}],[\"constructor\",{\"_index\":1,\"name\":{\"1\":{},\"26\":{},\"29\":{},\"32\":{},\"39\":{},\"50\":{},\"60\":{},\"66\":{},\"74\":{},\"78\":{},\"93\":{},\"103\":{},\"117\":{},\"129\":{},\"139\":{},\"150\":{},\"166\":{},\"174\":{},\"187\":{},\"206\":{},\"213\":{},\"223\":{},\"234\":{},\"259\":{},\"266\":{},\"300\":{},\"311\":{}},\"comment\":{}}],[\"conversion\",{\"_index\":43,\"name\":{\"53\":{}},\"comment\":{}}],[\"coordinatesystem\",{\"_index\":51,\"name\":{\"73\":{}},\"comment\":{}}],[\"copyto\",{\"_index\":8,\"name\":{\"8\":{}},\"comment\":{}}],[\"createdefaultfont\",{\"_index\":150,\"name\":{\"309\":{}},\"comment\":{}}],[\"dataset\",{\"_index\":27,\"name\":{\"31\":{}},\"comment\":{}}],[\"default\",{\"_index\":154,\"name\":{\"315\":{}},\"comment\":{}}],[\"destroy\",{\"_index\":7,\"name\":{\"7\":{}},\"comment\":{}}],[\"dispose\",{\"_index\":25,\"name\":{\"27\":{},\"30\":{},\"36\":{},\"86\":{},\"98\":{},\"110\":{},\"123\":{},\"134\":{},\"145\":{},\"161\":{},\"170\":{},\"181\":{},\"198\":{},\"209\":{},\"218\":{},\"229\":{},\"239\":{},\"261\":{},\"269\":{},\"308\":{},\"314\":{}},\"comment\":{}}],[\"disposed\",{\"_index\":29,\"name\":{\"34\":{}},\"comment\":{}}],[\"domain\",{\"_index\":37,\"name\":{\"44\":{},\"55\":{}},\"comment\":{}}],[\"font\",{\"_index\":151,\"name\":{\"310\":{}},\"comment\":{}}],[\"getnamedcommands\",{\"_index\":9,\"name\":{\"9\":{}},\"comment\":{}}],[\"grid\",{\"_index\":133,\"name\":{\"258\":{},\"260\":{}},\"comment\":{}}],[\"gridoptions\",{\"_index\":134,\"name\":{\"262\":{}},\"comment\":{}}],[\"height\",{\"_index\":17,\"name\":{\"18\":{},\"193\":{},\"306\":{}},\"comment\":{}}],[\"hlines\",{\"_index\":61,\"name\":{\"92\":{}},\"comment\":{}}],[\"hlinesoptions\",{\"_index\":65,\"name\":{\"99\":{}},\"comment\":{}}],[\"info\",{\"_index\":136,\"name\":{\"267\":{}},\"comment\":{}}],[\"instances\",{\"_index\":99,\"name\":{\"196\":{}},\"comment\":{}}],[\"interleavedcircles\",{\"_index\":66,\"name\":{\"102\":{}},\"comment\":{}}],[\"interleavedcirclesoptions\",{\"_index\":68,\"name\":{\"111\":{}},\"comment\":{}}],[\"interleavedshapes\",{\"_index\":69,\"name\":{\"116\":{}},\"comment\":{}}],[\"interleavedshapesoptions\",{\"_index\":73,\"name\":{\"124\":{}},\"comment\":{}}],[\"joincells\",{\"_index\":85,\"name\":{\"158\":{}},\"comment\":{}}],[\"joinids\",{\"_index\":84,\"name\":{\"157\":{}},\"comment\":{}}],[\"kerntable\",{\"_index\":152,\"name\":{\"312\":{}},\"comment\":{}}],[\"labelanchor\",{\"_index\":119,\"name\":{\"243\":{},\"281\":{}},\"comment\":{}}],[\"labelangle\",{\"_index\":120,\"name\":{\"244\":{},\"282\":{}},\"comment\":{}}],[\"labelcolor\",{\"_index\":121,\"name\":{\"245\":{},\"283\":{}},\"comment\":{}}],[\"labelformatter\",{\"_index\":141,\"name\":{\"274\":{}},\"comment\":{}}],[\"labelpad\",{\"_index\":122,\"name\":{\"246\":{},\"284\":{}},\"comment\":{}}],[\"labelside\",{\"_index\":123,\"name\":{\"247\":{},\"285\":{}},\"comment\":{}}],[\"labelsize\",{\"_index\":124,\"name\":{\"248\":{},\"286\":{}},\"comment\":{}}],[\"length\",{\"_index\":28,\"name\":{\"33\":{}},\"comment\":{}}],[\"linearscale\",{\"_index\":32,\"name\":{\"38\":{}},\"comment\":{}}],[\"lines\",{\"_index\":62,\"name\":{\"94\":{},\"214\":{}},\"comment\":{}}],[\"linesegments\",{\"_index\":74,\"name\":{\"128\":{}},\"comment\":{}}],[\"linesegmentsoptions\",{\"_index\":76,\"name\":{\"135\":{}},\"comment\":{}}],[\"linestrip\",{\"_index\":77,\"name\":{\"138\":{}},\"comment\":{}}],[\"linestripoptions\",{\"_index\":79,\"name\":{\"146\":{}},\"comment\":{}}],[\"logscale\",{\"_index\":42,\"name\":{\"49\":{}},\"comment\":{}}],[\"maxid\",{\"_index\":153,\"name\":{\"313\":{}},\"comment\":{}}],[\"minortickcolor\",{\"_index\":129,\"name\":{\"254\":{},\"292\":{}},\"comment\":{}}],[\"minortickcount\",{\"_index\":143,\"name\":{\"276\":{}},\"comment\":{}}],[\"minorticklength\",{\"_index\":130,\"name\":{\"255\":{},\"293\":{}},\"comment\":{}}],[\"minortickoffset\",{\"_index\":131,\"name\":{\"256\":{},\"294\":{}},\"comment\":{}}],[\"minorticks\",{\"_index\":115,\"name\":{\"238\":{},\"253\":{},\"291\":{},\"298\":{}},\"comment\":{}}],[\"minortickwidth\",{\"_index\":132,\"name\":{\"257\":{},\"295\":{}},\"comment\":{}}],[\"nameddrawcommands\",{\"_index\":23,\"name\":{\"24\":{}},\"comment\":{}}],[\"numberarray\",{\"_index\":19,\"name\":{\"20\":{}},\"comment\":{}}],[\"orthoaxis\",{\"_index\":135,\"name\":{\"265\":{}},\"comment\":{}}],[\"orthoaxisinfo\",{\"_index\":146,\"name\":{\"296\":{}},\"comment\":{}}],[\"orthoaxisoptions\",{\"_index\":137,\"name\":{\"270\":{}},\"comment\":{}}],[\"points\",{\"_index\":75,\"name\":{\"130\":{}},\"comment\":{}}],[\"polarcoordinatesystem\",{\"_index\":48,\"name\":{\"65\":{}},\"comment\":{}}],[\"position\",{\"_index\":94,\"name\":{\"188\":{}},\"comment\":{}}],[\"positionbuffer\",{\"_index\":59,\"name\":{\"85\":{},\"109\":{},\"169\":{},\"228\":{}},\"comment\":{}}],[\"primitive\",{\"_index\":24,\"name\":{\"25\":{}},\"comment\":{}}],[\"qdomain\",{\"_index\":35,\"name\":{\"42\":{}},\"comment\":{}}],[\"qrange\",{\"_index\":36,\"name\":{\"43\":{}},\"comment\":{}}],[\"quad\",{\"_index\":97,\"name\":{\"194\":{}},\"comment\":{}}],[\"quadgeometry\",{\"_index\":100,\"name\":{\"197\":{}},\"comment\":{}}],[\"radialscale\",{\"_index\":49,\"name\":{\"67\":{}},\"comment\":{}}],[\"radii\",{\"_index\":56,\"name\":{\"82\":{},\"89\":{},\"106\":{},\"114\":{},\"227\":{},\"232\":{}},\"comment\":{}}],[\"range\",{\"_index\":38,\"name\":{\"45\":{},\"56\":{}},\"comment\":{}}],[\"rects\",{\"_index\":89,\"name\":{\"165\":{},\"167\":{}},\"comment\":{}}],[\"rectsoptions\",{\"_index\":90,\"name\":{\"171\":{}},\"comment\":{}}],[\"recursiverender\",{\"_index\":10,\"name\":{\"10\":{}},\"comment\":{}}],[\"render\",{\"_index\":6,\"name\":{\"6\":{}},\"comment\":{}}],[\"renderable\",{\"_index\":18,\"name\":{\"19\":{}},\"comment\":{}}],[\"rotations\",{\"_index\":72,\"name\":{\"121\":{},\"127\":{},\"179\":{},\"185\":{}},\"comment\":{}}],[\"roundcapjoin\",{\"_index\":78,\"name\":{\"144\":{}},\"comment\":{}}],[\"scales\",{\"_index\":71,\"name\":{\"120\":{},\"126\":{},\"178\":{},\"184\":{}},\"comment\":{}}],[\"scissor\",{\"_index\":147,\"name\":{\"299\":{}},\"comment\":{}}],[\"scope\",{\"_index\":4,\"name\":{\"4\":{},\"302\":{}},\"comment\":{}}],[\"screenspace\",{\"_index\":149,\"name\":{\"307\":{}},\"comment\":{}}],[\"segmentcells\",{\"_index\":83,\"name\":{\"156\":{}},\"comment\":{}}],[\"segmentgeometry\",{\"_index\":64,\"name\":{\"97\":{},\"133\":{},\"217\":{}},\"comment\":{}}],[\"segmentpositions\",{\"_index\":82,\"name\":{\"155\":{}},\"comment\":{}}],[\"shape\",{\"_index\":70,\"name\":{\"118\":{},\"175\":{}},\"comment\":{}}],[\"shapes\",{\"_index\":91,\"name\":{\"173\":{}},\"comment\":{}}],[\"shapesoptions\",{\"_index\":92,\"name\":{\"182\":{}},\"comment\":{}}],[\"size\",{\"_index\":95,\"name\":{\"189\":{},\"204\":{}},\"comment\":{}}],[\"text\",{\"_index\":93,\"name\":{\"186\":{}},\"comment\":{}}],[\"textoptions\",{\"_index\":101,\"name\":{\"199\":{}},\"comment\":{}}],[\"texts\",{\"_index\":113,\"name\":{\"236\":{}},\"comment\":{}}],[\"tickcolor\",{\"_index\":125,\"name\":{\"249\":{},\"287\":{}},\"comment\":{}}],[\"ticklength\",{\"_index\":126,\"name\":{\"250\":{},\"288\":{}},\"comment\":{}}],[\"tickoffset\",{\"_index\":127,\"name\":{\"251\":{},\"289\":{}},\"comment\":{}}],[\"tickorigin\",{\"_index\":144,\"name\":{\"277\":{}},\"comment\":{}}],[\"ticks\",{\"_index\":114,\"name\":{\"237\":{},\"297\":{}},\"comment\":{}}],[\"tickstep\",{\"_index\":145,\"name\":{\"278\":{}},\"comment\":{}}],[\"tickwidth\",{\"_index\":128,\"name\":{\"252\":{},\"290\":{}},\"comment\":{}}],[\"todomain\",{\"_index\":39,\"name\":{\"46\":{},\"58\":{},\"64\":{},\"72\":{},\"75\":{}},\"comment\":{}}],[\"torange\",{\"_index\":40,\"name\":{\"47\":{},\"57\":{},\"63\":{},\"71\":{},\"76\":{}},\"comment\":{}}],[\"transparentlinestrip\",{\"_index\":80,\"name\":{\"149\":{}},\"comment\":{}}],[\"transparentlinestripoptions\",{\"_index\":88,\"name\":{\"162\":{}},\"comment\":{}}],[\"triangles\",{\"_index\":104,\"name\":{\"205\":{}},\"comment\":{}}],[\"trianglesoptions\",{\"_index\":106,\"name\":{\"210\":{}},\"comment\":{}}],[\"update\",{\"_index\":30,\"name\":{\"35\":{}},\"comment\":{}}],[\"updateq\",{\"_index\":41,\"name\":{\"48\":{}},\"comment\":{}}],[\"uv\",{\"_index\":98,\"name\":{\"195\":{}},\"comment\":{}}],[\"vector2\",{\"_index\":20,\"name\":{\"21\":{}},\"comment\":{}}],[\"vector3\",{\"_index\":21,\"name\":{\"22\":{}},\"comment\":{}}],[\"vector4\",{\"_index\":22,\"name\":{\"23\":{}},\"comment\":{}}],[\"vertices\",{\"_index\":105,\"name\":{\"207\":{}},\"comment\":{}}],[\"viewport\",{\"_index\":13,\"name\":{\"14\":{}},\"comment\":{}}],[\"vlines\",{\"_index\":107,\"name\":{\"212\":{}},\"comment\":{}}],[\"vlinesoptions\",{\"_index\":108,\"name\":{\"219\":{}},\"comment\":{}}],[\"wedges\",{\"_index\":109,\"name\":{\"222\":{}},\"comment\":{}}],[\"wedgesoptions\",{\"_index\":111,\"name\":{\"230\":{}},\"comment\":{}}],[\"width\",{\"_index\":16,\"name\":{\"17\":{},\"153\":{},\"163\":{},\"192\":{},\"263\":{},\"305\":{}},\"comment\":{}}],[\"widths\",{\"_index\":63,\"name\":{\"95\":{},\"100\":{},\"131\":{},\"136\":{},\"142\":{},\"147\":{},\"215\":{},\"220\":{}},\"comment\":{}}],[\"x\",{\"_index\":14,\"name\":{\"15\":{},\"303\":{}},\"comment\":{}}],[\"xs\",{\"_index\":53,\"name\":{\"79\":{},\"140\":{},\"151\":{},\"176\":{}},\"comment\":{}}],[\"xscale\",{\"_index\":46,\"name\":{\"61\":{},\"69\":{}},\"comment\":{}}],[\"xys\",{\"_index\":67,\"name\":{\"104\":{},\"119\":{},\"224\":{}},\"comment\":{}}],[\"y\",{\"_index\":15,\"name\":{\"16\":{},\"304\":{}},\"comment\":{}}],[\"ys\",{\"_index\":54,\"name\":{\"80\":{},\"141\":{},\"152\":{},\"177\":{}},\"comment\":{}}],[\"yscale\",{\"_index\":47,\"name\":{\"62\":{},\"70\":{}},\"comment\":{}}]],\"pipeline\":[]}}");
