+++ docs/api/classes/Scissor.html
@@ -95,38 +95,38 @@
 <h3 class="tsd-anchor-link"><code class="tsd-tag ts-flagPrivate">Private</code> <span>_children</span><a href="#_children" aria-label="Permalink" class="tsd-anchor-icon"><svg class="icon icon-tabler icon-tabler-link" viewBox="0 0 24 24" stroke-width="2" stroke="currentColor" fill="none" stroke-linecap="round" stroke-linejoin="round"><use href="#icon-anchor-a"></use><use href="#icon-anchor-b"></use><use href="#icon-anchor-c"></use></svg></a></h3>
 <div class="tsd-signature">_children<span class="tsd-signature-symbol">:</span> <a href="../types/Renderable.html" class="tsd-signature-type" data-tsd-kind="Type alias">Renderable</a></div><aside class="tsd-sources">
 <ul>
-<li>Defined in <a href="https://github.com/wwwtyro/candygraph/blob/c1396ff/src/composites/scissor.ts#L42">composites/scissor.ts:42</a></li></ul></aside></section>
+<li>Defined in <a href="https://github.com/wwwtyro/candygraph/blob/c1396ff/src/composites/scissor.ts#L43">composites/scissor.ts:43</a></li></ul></aside></section>
 <section class="tsd-panel tsd-member tsd-kind-property tsd-parent-kind-class"><a id="height" class="tsd-anchor"></a>
 <h3 class="tsd-anchor-link"><span>height</span><a href="#height" aria-label="Permalink" class="tsd-anchor-icon"><svg class="icon icon-tabler icon-tabler-link" viewBox="0 0 24 24" stroke-width="2" stroke="currentColor" fill="none" stroke-linecap="round" stroke-linejoin="round"><use href="#icon-anchor-a"></use><use href="#icon-anchor-b"></use><use href="#icon-anchor-c"></use></svg></a></h3>
 <div class="tsd-signature">height<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">number</span></div><aside class="tsd-sources">
 <ul>
-<li>Defined in <a href="https://github.com/wwwtyro/candygraph/blob/c1396ff/src/composites/scissor.ts#L58">composites/scissor.ts:58</a></li></ul></aside></section>
+<li>Defined in <a href="https://github.com/wwwtyro/candygraph/blob/c1396ff/src/composites/scissor.ts#L59">composites/scissor.ts:59</a></li></ul></aside></section>
 <section class="tsd-panel tsd-member tsd-kind-property tsd-parent-kind-class"><a id="scope" class="tsd-anchor"></a>
 <h3 class="tsd-anchor-link"><code class="tsd-tag ts-flagReadonly">Readonly</code> <span>scope</span><a href="#scope" aria-label="Permalink" class="tsd-anchor-icon"><svg class="icon icon-tabler icon-tabler-link" viewBox="0 0 24 24" stroke-width="2" stroke="currentColor" fill="none" stroke-linecap="round" stroke-linejoin="round"><use href="#icon-anchor-a"></use><use href="#icon-anchor-b"></use><use href="#icon-anchor-c"></use></svg></a></h3>
 <div class="tsd-signature">scope<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">DrawCommand</span><span class="tsd-signature-symbol">&lt;</span><span class="tsd-signature-type">DefaultContext</span><span class="tsd-signature-symbol">, </span><span class="tsd-signature-symbol">{}</span><span class="tsd-signature-symbol">&gt;</span></div><aside class="tsd-sources">
 <p>Overrides Composite.scope</p>
 <ul>
-<li>Defined in <a href="https://github.com/wwwtyro/candygraph/blob/c1396ff/src/composites/scissor.ts#L43">composites/scissor.ts:43</a></li></ul></aside></section>
+<li>Defined in <a href="https://github.com/wwwtyro/candygraph/blob/c1396ff/src/composites/scissor.ts#L44">composites/scissor.ts:44</a></li></ul></aside></section>
 <section class="tsd-panel tsd-member tsd-kind-property tsd-parent-kind-class"><a id="screenSpace" class="tsd-anchor"></a>
 <h3 class="tsd-anchor-link"><span>screen<wbr/>Space</span><a href="#screenSpace" aria-label="Permalink" class="tsd-anchor-icon"><svg class="icon icon-tabler icon-tabler-link" viewBox="0 0 24 24" stroke-width="2" stroke="currentColor" fill="none" stroke-linecap="round" stroke-linejoin="round"><use href="#icon-anchor-a"></use><use href="#icon-anchor-b"></use><use href="#icon-anchor-c"></use></svg></a></h3>
 <div class="tsd-signature">screen<wbr/>Space<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">boolean</span></div><aside class="tsd-sources">
 <ul>
-<li>Defined in <a href="https://github.com/wwwtyro/candygraph/blob/c1396ff/src/composites/scissor.ts#L59">composites/scissor.ts:59</a></li></ul></aside></section>
+<li>Defined in <a href="https://github.com/wwwtyro/candygraph/blob/c1396ff/src/composites/scissor.ts#L60">composites/scissor.ts:60</a></li></ul></aside></section>
 <section class="tsd-panel tsd-member tsd-kind-property tsd-parent-kind-class"><a id="width" class="tsd-anchor"></a>
 <h3 class="tsd-anchor-link"><span>width</span><a href="#width" aria-label="Permalink" class="tsd-anchor-icon"><svg class="icon icon-tabler icon-tabler-link" viewBox="0 0 24 24" stroke-width="2" stroke="currentColor" fill="none" stroke-linecap="round" stroke-linejoin="round"><use href="#icon-anchor-a"></use><use href="#icon-anchor-b"></use><use href="#icon-anchor-c"></use></svg></a></h3>
 <div class="tsd-signature">width<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">number</span></div><aside class="tsd-sources">
 <ul>
-<li>Defined in <a href="https://github.com/wwwtyro/candygraph/blob/c1396ff/src/composites/scissor.ts#L57">composites/scissor.ts:57</a></li></ul></aside></section>
+<li>Defined in <a href="https://github.com/wwwtyro/candygraph/blob/c1396ff/src/composites/scissor.ts#L58">composites/scissor.ts:58</a></li></ul></aside></section>
 <section class="tsd-panel tsd-member tsd-kind-property tsd-parent-kind-class"><a id="x" class="tsd-anchor"></a>
 <h3 class="tsd-anchor-link"><span>x</span><a href="#x" aria-label="Permalink" class="tsd-anchor-icon"><svg class="icon icon-tabler icon-tabler-link" viewBox="0 0 24 24" stroke-width="2" stroke="currentColor" fill="none" stroke-linecap="round" stroke-linejoin="round"><use href="#icon-anchor-a"></use><use href="#icon-anchor-b"></use><use href="#icon-anchor-c"></use></svg></a></h3>
 <div class="tsd-signature">x<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">number</span></div><aside class="tsd-sources">
 <ul>
-<li>Defined in <a href="https://github.com/wwwtyro/candygraph/blob/c1396ff/src/composites/scissor.ts#L55">composites/scissor.ts:55</a></li></ul></aside></section>
+<li>Defined in <a href="https://github.com/wwwtyro/candygraph/blob/c1396ff/src/composites/scissor.ts#L56">composites/scissor.ts:56</a></li></ul></aside></section>
 <section class="tsd-panel tsd-member tsd-kind-property tsd-parent-kind-class"><a id="y" class="tsd-anchor"></a>
 <h3 class="tsd-anchor-link"><span>y</span><a href="#y" aria-label="Permalink" class="tsd-anchor-icon"><svg class="icon icon-tabler icon-tabler-link" viewBox="0 0 24 24" stroke-width="2" stroke="currentColor" fill="none" stroke-linecap="round" stroke-linejoin="round"><use href="#icon-anchor-a"></use><use href="#icon-anchor-b"></use><use href="#icon-anchor-c"></use></svg></a></h3>
 <div class="tsd-signature">y<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">number</span></div><aside class="tsd-sources">
 <ul>
-<li>Defined in <a href="https://github.com/wwwtyro/candygraph/blob/c1396ff/src/composites/scissor.ts#L56">composites/scissor.ts:56</a></li></ul></aside></section></section>
+<li>Defined in <a href="https://github.com/wwwtyro/candygraph/blob/c1396ff/src/composites/scissor.ts#L57">composites/scissor.ts:57</a></li></ul></aside></section></section>
 <section class="tsd-panel-group tsd-member-group">
 <h2>Methods</h2>
 <section class="tsd-panel tsd-member tsd-kind-method tsd-parent-kind-class tsd-is-inherited"><a id="dispose" class="tsd-anchor"></a>
