+++ docs/api/types/NumberArray.html
@@ -15,7 +15,7 @@
 <h1>Type alias NumberArray</h1></div>
 <div class="tsd-signature">Number<wbr/>Array<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">number</span><span class="tsd-signature-symbol">[]</span><span class="tsd-signature-symbol"> | </span><span class="tsd-signature-type">Float32Array</span></div><aside class="tsd-sources">
 <ul>
-<li>Defined in <a href="https://github.com/wwwtyro/candygraph/blob/c1396ff/src/common.ts#L53">common.ts:53</a></li></ul></aside></div>
+<li>Defined in <a href="https://github.com/wwwtyro/candygraph/blob/c1396ff/src/common.ts#L20">common.ts:20</a></li></ul></aside></div>
 <div class="col-4 col-menu menu-sticky-wrap menu-highlight">
 <div class="tsd-navigation settings">
 <details class="tsd-index-accordion"><summary class="tsd-accordion-summary">
