+++ docs/api/types/Renderable.html
@@ -15,7 +15,7 @@
 <h1>Type alias Renderable</h1></div>
 <div class="tsd-signature">Renderable<span class="tsd-signature-symbol">:</span> <a href="../classes/Primitive.html" class="tsd-signature-type" data-tsd-kind="Class">Primitive</a><span class="tsd-signature-symbol"> | </span><a href="../classes/Composite.html" class="tsd-signature-type" data-tsd-kind="Class">Composite</a><span class="tsd-signature-symbol"> | </span><a href="Renderable.html" class="tsd-signature-type" data-tsd-kind="Type alias">Renderable</a><span class="tsd-signature-symbol">[]</span></div><aside class="tsd-sources">
 <ul>
-<li>Defined in <a href="https://github.com/wwwtyro/candygraph/blob/c1396ff/src/common.ts#L51">common.ts:51</a></li></ul></aside></div>
+<li>Defined in <a href="https://github.com/wwwtyro/candygraph/blob/c1396ff/src/common.ts#L18">common.ts:18</a></li></ul></aside></div>
 <div class="col-4 col-menu menu-sticky-wrap menu-highlight">
 <div class="tsd-navigation settings">
 <details class="tsd-index-accordion"><summary class="tsd-accordion-summary">
