+++ docs/api/types/Vector2.html
@@ -15,7 +15,7 @@
 <h1>Type alias Vector2</h1></div>
 <div class="tsd-signature">Vector2<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">number</span><span class="tsd-signature-symbol">[]</span><span class="tsd-signature-symbol"> | </span><span class="tsd-signature-type">Float32Array</span></div><aside class="tsd-sources">
 <ul>
-<li>Defined in <a href="https://github.com/wwwtyro/candygraph/blob/c1396ff/src/common.ts#L58">common.ts:58</a></li></ul></aside></div>
+<li>Defined in <a href="https://github.com/wwwtyro/candygraph/blob/c1396ff/src/common.ts#L25">common.ts:25</a></li></ul></aside></div>
 <div class="col-4 col-menu menu-sticky-wrap menu-highlight">
 <div class="tsd-navigation settings">
 <details class="tsd-index-accordion"><summary class="tsd-accordion-summary">
