+++ docs/api/types/Vector3.html
@@ -15,7 +15,7 @@
 <h1>Type alias Vector3</h1></div>
 <div class="tsd-signature">Vector3<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">number</span><span class="tsd-signature-symbol">[]</span><span class="tsd-signature-symbol"> | </span><span class="tsd-signature-type">Float32Array</span></div><aside class="tsd-sources">
 <ul>
-<li>Defined in <a href="https://github.com/wwwtyro/candygraph/blob/c1396ff/src/common.ts#L59">common.ts:59</a></li></ul></aside></div>
+<li>Defined in <a href="https://github.com/wwwtyro/candygraph/blob/c1396ff/src/common.ts#L26">common.ts:26</a></li></ul></aside></div>
 <div class="col-4 col-menu menu-sticky-wrap menu-highlight">
 <div class="tsd-navigation settings">
 <details class="tsd-index-accordion"><summary class="tsd-accordion-summary">
