+++ src/common.ts
@@ -1,6 +1,5 @@
-import { DrawCommand } from "regl";
-import { CoordinateSystem } from "./coordinates/coordinate-system";
 import { Primitive } from "./primitives/primitive";
+import { Composite } from "./composites/composite";
 
 export type UnpackPromise<T> = T extends Promise<infer U> ? U : never;
 
@@ -14,38 +13,6 @@ export interface Viewport {
 export enum RenderableType {
   Primitive,
   Composite,
-}
-
-export abstract class Composite {
-  /** @internal */
-  public readonly kind = RenderableType.Composite;
-
-  /** @internal */
-  public abstract children(): Renderable;
-
-  /** @internal */
-  public readonly scope: DrawCommand | null = null;
-
-  /** @internal */
-  public props(coords: CoordinateSystem) {
-    return {};
-  }
-
-  /** Releases all GPU resources and renders this instance unusable. Also disposes any children. */
-  public dispose(): void {
-    function recurse(renderable: Renderable) {
-      if (Array.isArray(renderable)) {
-        for (const element of renderable) {
-          recurse(element);
-        }
-      } else if (renderable.kind === RenderableType.Composite) {
-        recurse(renderable.children());
-      } else if (renderable.kind === RenderableType.Primitive) {
-        renderable.dispose();
-      }
-    }
-    recurse(this.children());
-  }
 }
 
 export type Renderable = Primitive | Composite | Renderable[];
