+++ src/composites/axis.ts
@@ -1,7 +1,8 @@
 import { vec2 } from "gl-matrix";
 import { CandyGraph } from "../candygraph";
 import { CoordinateSystem } from "../coordinates/coordinate-system";
-import { Composite, Renderable, Vector4, Vector2, NumberArray } from "../common";
+import { Renderable, Vector4, Vector2, NumberArray } from "../common";
+import { Composite } from "./composite";
 import { Text } from "../primitives/text";
 import { VLines } from "../primitives/vlines";
 import { HLines } from "../primitives/hlines";
