+++ src/index.ts
@@ -3,9 +3,11 @@ export default CandyGraph;
 
 export { CandyGraph, CandyGraphOptions } from "./candygraph";
 
-export { Viewport, Composite, Renderable, NumberArray, Vector2, Vector3, Vector4 } from "./common";
+export { Viewport, Renderable, NumberArray, Vector2, Vector3, Vector4 } from "./common";
 
 export { NamedDrawCommands, Primitive } from "./primitives/primitive";
+
+export { Composite } from "./composites/composite";
 
 export { Dataset } from "./dataset";
